### 2023-05-04 #5441 @editorialbot
[PRE REVIEW]: nctoolkit: A Python package for netCDF analysis and post-processing
**Submitting author:** @robertjwilson (<a href="http://orcid.org/0000-0002-0592-366X">Robert Wilson</a>)
**Repository:** https://github.com/pmlmodelling/nctoolkit
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.2
