+++ requirements.txt
@@ -8,3 +8,4 @@ dill
 python-dateutil
 matplotlib
 
+tornado>=6.3.2 # not directly required, pinned by Snyk to avoid a vulnerability
