+++ docs/requirements.txt
@@ -16,3 +16,4 @@ dill
 
 
 
+tornado>=6.3.2 # not directly required, pinned by Snyk to avoid a vulnerability
