### 2023-05-26 #5441 @platipodium
@martinfleis I work a lot with netcdf files, so if you agree, I'd offer myself as a reviewer for this submission.
