### 2023-05-27 pmlmodelling/nctoolkit#57 @platipodium
[JOSS] statement of need - dask
For your statement of need

> For example, the Coupled Model Intercomparison Project Phase 6 (O’Neill et al., 2016), produced approximately 20 PB of publicly available data (Petrie et al., 2021).

Please clarify why you would not rely on technologies like `dask` that would distribute the work, especially with those large data files.
