### 2023-05-27 #5494 @platipodium
Explain why not use dask or similar tech: https://github.com/pmlmodelling/nctoolkit/issues/57
