### 2023-05-27 pmlmodelling/nctoolkit#58 @platipodium
[JOSS] statement of need - compare to other software
In your statement of need you write

> It provides a clean interface for working with netCDF output, which often have irregular vertical grids.

This is indeed sufficient explanation for needing this particular software.  We have, however other challenges working with netCDF that are addressed by other packages, e.g. psyplot https://www.theoj.org/joss-papers/joss.00363/10.21105.joss.00363.pdf targeting unstructured grids, or Thalassa https://github.com/ec-jrc/Thalassa for large-scale seamless visualisation.

It would be great if you did some research on other netCDF science viz solutions, and contrast them in your JOSS paper, such that the added value of your specific solution is highlighted.
