### 2023-05-27 #5494 @platipodium
Discriminate from other software capabilities: https://github.com/pmlmodelling/nctoolkit/issues/58
