### 2023-05-27 pmlmodelling/nctoolkit#59 @platipodium
[JOSS] Summary target lay audience better
In your summary section, make sure that it is understandable to people who do not know what netCDF is.  This need only a little elaboration.
