### 2023-05-27 #5494 @platipodium
Portability problems to Mac running Monterey/Ventura python 3.10 + 3.11, see https://github.com/pmlmodelling/nctoolkit/issues/29
