### 2023-05-27 pmlmodelling/nctoolkit#60 @platipodium
[JOSS] Missing declaration of optional dependency on cartopy
Neither `conda` nor `pip` install `cartopy` as a dependency.  This leads to garbled plots on my machine, but you do provide a helpful message

> Unable to import cartopy. For better plots install cartopy or check cartopy installation!

Maybe this should be documented.

https://github.com/openjournals/joss-reviews/issues/5494
