### 2023-05-27 #5494 @platipodium
Documentation of cartopy dependency missing: https://github.com/pmlmodelling/nctoolkit/issues/60
