### 2023-05-27 pmlmodelling/nctoolkit#61 @platipodium
Garbled plot
Running the example

```
import nctoolkit as nc
ds = nc.open_data("sst.mon.mean.nc")
ds.plot()
```

Creates a garbled plot on my machine (Mac M1 mamba/pip python11), problem in all browsers (Safari, Firefox, Chrome).  Y dimension not correct

https://github.com/openjournals/joss-reviews/issues/5494
