### 2023-05-27 #5494 @platipodium
Garbled plot from example: https://github.com/pmlmodelling/nctoolkit/issues/61
