+++ .circleci/config.yml
@@ -2,6 +2,10 @@
 version: 2.1
 orbs:
   codecov: codecov/codecov@3.1.1
+
+macos:
+  xcode: 14.2.0 # indicate your selected version of Xcode
+resource_class: macos.m1.large.gen1
 jobs:
   build:
     docker:
