+++ .circleci/config.yml
@@ -3,11 +3,11 @@ version: 2.1
 orbs:
   codecov: codecov/codecov@3.1.1
 
-macos:
-  xcode: 14.2.0 # indicate your selected version of Xcode
-resource_class: macos.m1.large.gen1
 jobs:
   build:
+    macos:
+      xcode: 14.2.0 # indicate your selected version of Xcode
+    resource_class: macos.m1.large.gen1
     docker:
       - image: continuumio/miniconda3
 
