+++ .circleci/config.yml
@@ -5,10 +5,10 @@ orbs:
 
 jobs:
   build:
-    macos:
-      xcode: 14.2.0 # indicate your selected version of Xcode
-    resource_class: macos.m1.large.gen1
     docker:
+      macos:
+        xcode: 14.2.0 # indicate your selected version of Xcode
+      resource_class: macos.m1.large.gen1
       - image: continuumio/miniconda3
 
     working_directory: ~/repo
