### 2023-06-01 pmlmodelling/nctoolkit#61 @robertjwilson
I am able to reproduce this problem @platipodium . It appears to be caused by an update to panel 14 days ago. Downgrading panel to 0.14.4 fixes the issue.

I'll pin the panel version on conda and do a quick release on pypi to pin it there. Meanwhile, I'll investigate whether the nctoolkit plotting code needs to change or whether it's something hvplot devs need to fix.
