### 2023-06-02 pmlmodelling/nctoolkit#61 @robertjwilson
This should now be fixed temporarily with a quick release I've just made to nctoolkit's plotting dependency ncplot: [here](https://github.com/pmlmodelling/ncplot/releases/tag/v0.3.3).

The cause of it was that hvplot's `image` method was not able to fit the browser window when you set `responsive=True`. I've fixed the problem by setting `responsive=False`. This means you can get a smallish plot window, but it's better than something compressed.

However, they appear to be on top of the issue at hvplot. I pointed it out to them and they said a code update was coming in the next day or so and a new release fixing the problem is incoming: [here](https://github.com/holoviz/hvplot/issues/1074#issuecomment-1573323252)
