### 2023-06-02 pmlmodelling/nctoolkit#62 @platipodium
[JOSS] Add guidelines on how to contribute
Especially clarify

- the desired workflow for contributions
- the legal ramifications (Contributor license agreement) for contributions

https://github.com/openjournals/joss-reviews/issues/5494
