### 2023-06-02 pmlmodelling/nctoolkit#63 @platipodium
[JOSS] Add guideline for how to seek support / report problems
Especially clarify

- the desired communication path (email,  issue, wiki, forum ...)  for support
- the desired communication path (email,  issue, wiki, forum ...)  for  problems

https://github.com/openjournals/joss-reviews/issues/5494
