### 2023-06-02 #5494 @platipodium
For better community integration, please consider the following:

- https://github.com/pmlmodelling/nctoolkit/issues/62
- https://github.com/pmlmodelling/nctoolkit/issues/63
