### 2023-06-05 ecf8c0f
add python scripting test
