### 2023-06-05 c9de76f
add python scripting test
