+++ .circleci/config.yml
@@ -6,11 +6,11 @@ orbs:
 
 jobs:
   build:
-    macos:
-      xcode: 14.2.0 # indicate your selected version of Xcode
-    resource_class: macos.m1.large.gen1
-    #docker:
-    #  - image: continuumio/miniconda3
+    #macos:
+    #  xcode: 14.2.0 # indicate your selected version of Xcode
+    #resource_class: macos.m1.large.gen1
+    docker:
+      - image: continuumio/miniconda3
 
     working_directory: ~/repo
 
