### 2023-06-05 a504c84
run python script test on osx
