+++ testing/test_scripting.py
@@ -1,0 +1,5 @@
+import nctoolkit as nc
+
+ds = nc.open_data("data/sst.mon.mean.nc")
+ds.tmean()
+ds.run()
