### 2023-06-05 a2a0141
switch to multiprocess for list
