### 2023-06-05 6ea69a2
add tests back in for macos
