### 2023-06-05 103e858
use multiprocessing for lists in linux
