### 2023-06-05 bd2e0e3
add test for running nctoolkit in parallel
