### 2023-06-05 bf836ee
drop mp tests for formal tests
