+++ nctoolkit/cleanup.py
@@ -62,16 +62,9 @@ def cleanup():
     if len(delete_these) == 0:
         return None
 
-    # ignore stuff less than a minute old...
-
-    if session_info["parallel"]:
-        for dd in delete_these:
-            if os.path.exists(dd) and file_age_in_seconds(dd) > 60:
-                nc_remove(dd)
-    else:
-        for dd in delete_these:
-            if os.path.exists(dd):
-                nc_remove(dd)
+    for dd in delete_these:
+        if os.path.exists(dd):
+            nc_remove(dd)
 
     # only update the session size on linux
 
