### 2023-06-06 pmlmodelling/nctoolkit#65 @platipodium
[JOSS] ambiguous data path in paper example
In the JOSS paper, you write

```
ds_clim = nc.open_data("sst.mon.mean.nc")
ds_present = nc.open_data("data/sst.mon.mean.nc")
```

This is misleading.  Are both datasets the same file?  Why in different folders?

See https://github.com/openjournals/joss-reviews/issues/5494
