### 2023-06-06 09ff734
Update paper.md
Fix typo in file path
