### 2023-06-06 pmlmodelling/nctoolkit#66 @platipodium
[JOSS] Dependency ncplot is not well documented
The dependency `ncplot` is not very well documented.  Any suggestions  or plans to improve that?

https://ncplot.readthedocs.io/en/latest/index.html

See https://github.com/openjournals/joss-reviews/issues/5494
