### 2023-06-06 pmlmodelling/nctoolkit#66 @robertjwilson
I've just checked this. For some reason readthedocs was disliking a notebook. I've replaced the .ipynb file with a .rst file.

https://ncplot.readthedocs.io/en/latest/index.html
