### 2023-06-06 pmlmodelling/nctoolkit#66 @platipodium
OK, I see.  There is indeed only *one* API call: `view`.  That simple :=).  I guess I was missing information on how to style the plot once it is done with `view`.  That should be documented

- what kind of plot object is created?
- with what API is the plot object to be manipulated?

That's at least what I was wondering when I played around with using nctoolkit to viz some of my own files.
