### 2023-06-07 f30a2cf
minor improvement to docstring
