### 2023-06-07 pmlmodelling/nctoolkit#66 @platipodium
I think your comment in the docs

> This is a package that aims to deliver plotting for rapid exploratory analysis, and therefore it does not offer a large number of customizations. However, because it is built on hvplot, you can use most of the customization options available in hvplot, which are detailed [here](https://hvplot.holoviz.org/user_guide/Customization.html). Arguments such as logz and clim can be passed to plot and will be automatically passed to the hvplot method used.

is exactly what I was looking for.   Thanks!

Now, however, the readthedocs is broken.  I'll create a new issue
