### 2023-06-07 pmlmodelling/nctoolkit#67 @platipodium
[JOSS] readthedocs is broken
The readthedos at https://nctoolkit.readthedocs.io/en/latest/ is broken.

https://github.com/openjournals/joss-reviews/issues/5494
