+++ docs/source/conf.py
@@ -39,7 +39,6 @@ extensions = [
 ]
 
 html_theme = "sphinx_rtd_theme"
-#html_theme = "haiku"
 html_static_path = ['_static']
 html_logo = "_static/nctoolkit_logo.jpg"
 
