### 2023-06-07 pmlmodelling/nctoolkit#67 @robertjwilson
The docs should be fine now. I think it was broken because it was in the middle of a re-build.
