+++ docs/source/visualization_1.rst
@@ -22,9 +22,8 @@ Let’s start by calculating mean sea surface temperature for the year
     ds.plot()
 
 .. raw:: html
-   :file: visualization_plot1.html 
 
-
+   <iframe src="visualization_plot1.html" width="100%" height="350px"></iframe>
 
 
 We might be interested in the zonal mean. nctoolkit can automatically
