### 2023-06-07 pmlmodelling/nctoolkit#68 @malmans2
[JOSS] Compare nctoolkit with existing tools and find a better example
Hi there,

Nice package! I didn't have the time to review it yet, but I glanced through the documentation and the paper.

My first comment is actually a question that I think you should address both in the paper and in the documentation: How does this tool compare to other existing packages?

I don't find the example in the paper particularly convincing, so my suggestion is to find a better one that really highlights the strengths of `nctoolkit`.

`xarray` is my go-to library for this kind of processing, and the example in the paper looks quite similar (although the plot is not publication ready, so that's definitely a strength of nctoolkit):
```python
import fsspec
import cartopy.crs as ccrs
import xarray as xr

url = "https://downloads.psl.noaa.gov/Datasets/COBE/sst.mon.mean.nc"
with fsspec.open(f"simplecache::{url}", simplecache={"same_names": True}) as f:
    ds = xr.open_mfdataset(f.name)

ds_clim = ds.sel(time=slice("1900", "1919")).mean("time")
ds_pres = ds.sel(time=slice("2000", "2019")).mean("time")
ds_anom = ds_pres - ds_clim

ds_anom["sst"].plot(
    transform=ccrs.PlateCarree(), subplot_kws={"projection": ccrs.Robinson()}
)
```
![image](https://github.com/pmlmodelling/nctoolkit/assets/22245117/7b5dd23e-0736-46dd-bbe5-a3c769400bcc)

https://github.com/openjournals/joss-reviews/issues/5494
