### 2023-06-07 pmlmodelling/nctoolkit#69 @malmans2
[JOSS] Clarify which libraries are used for core functionalities
The paper mentions that `nctoolkit` is built on top of CDO.
Therefore, I wouldn't expect to see that many libraries in the [requirement list](https://github.com/pmlmodelling/nctoolkit/blob/master/requirements.txt).

I'll use `xarray` as example. Is it only used to return `xarray` objects?
If that's the case, I would make it optional and return an informative error if `to_xarray()` is used and `xarray` is not installed.

https://github.com/openjournals/joss-reviews/issues/5494
