### 2023-06-07 pmlmodelling/nctoolkit#69 @malmans2
Got it. Because plotting is just one of many core functionalities, my suggestion is to keep the list of mandatory dependencies small.

It looks like you already isolated plotting in another package ([ncplot](https://github.com/pmlmodelling/ncplot)).

How about:
```
pip install nctoolkit  # minimal
pip install nctoolkit[plot]  # plotting packages as well
pip install nctoolkit[complete] # everything
```

With conda:
```
conda install -c conda-forge nctoolkit  # minimal
```
or
```
conda install -c conda-forge nctoolkit ncplot  # plotting packages as well
```
