### 2023-06-07 pmlmodelling/nctoolkit#70 @malmans2
[JOSS] Speed up tests and address warnings
I was looking at the GH actions, and I noticed it takes +1h to run the tests.
It would be nice to speed up both the setup (+10min) and testing steps (+30min), especially if you encourage external contributions.

For the setup step, I recommend to use this action: https://github.com/mamba-org/setup-micromamba
Micromamba is generally much faster, and the action provides an option to activate the environment cache.

Also, if you didn't notice it, there are a few warnings raised both by `pytest` and GitHub actions.
For example, see: https://github.com/pmlmodelling/nctoolkit/actions/runs/5198818724
I suggest to solve the GitHub actions warnings, and [assert](https://docs.pytest.org/en/7.1.x/how-to/capture-warnings.html#warns) or solve pytest warnings.

https://github.com/openjournals/joss-reviews/issues/5494
