+++ .github/workflows/python-app-linux.yml
@@ -16,7 +16,8 @@ jobs:
 
     steps:
 
-     - uses: goanpeca/setup-miniconda@v1
+     #- uses: goanpeca/setup-miniconda@v1
+     - uses: mamba-org/setup-micromamba@v1 
        env:
         ACTIONS_ALLOW_UNSECURE_COMMANDS: true
        with:
@@ -38,15 +39,11 @@ jobs:
      - name: Install dependencies
        shell: bash -l {0}
        run: |
-         conda activate foo
-         conda install cdo pip nco geos -y
+         mamba activate foo
+         mamba install cdo pip nco geos -y
          pip install numpy 
          pip install cartopy
          pip install dask
-         #conda install numpy -y
-         #conda install cartopy -y
-         #conda install dask -y
-         #    conda install flake8
          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
          pip install .
          pip install pandas
@@ -55,18 +52,10 @@ jobs:
          pip install pytest-cov
          pip install coverage 
 
-         #- name: Lint with flake8
-         #  shell: bash -l {0}
-         #  run: |
-         #    conda activate foo
-         #    # stop the build if there are Python syntax errors or undefined names
-         #    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-         #    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-         #    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
      - name: Test with pytest
        shell: bash -l {0}
        run: |
-         conda activate foo
+         mamba activate foo
          #pytest tests/test*.py
          python testing/test_parallel.py
          coverage run -m pytest tests/test*.py
