### 2023-06-07 4a689b8
strip unnedded tests
