+++ .github/workflows/python-app-mamba.yml
@@ -25,13 +25,7 @@ jobs:
             powershell
           cache-environment: true
           post-cleanup: 'all'
-      - name: Import numpy in micromamba environment (bash)
-        run: python -c "import numpy"
-        shell: bash -el {0}
-      - name: Import numpy in micromamba environment (pwsh)
-        run: python -c "import numpy"
-        shell: pwsh
-      - name: Run custom command in micromamba environment
+      - name: Run python testing
         run: 
           mamba activate foo
           mamba install cdo nco
