### 2023-06-07 f633945
pull in environment for mamba test
