+++ .github/workflows/python-app-mamba.yml
@@ -19,7 +19,7 @@ jobs:
       - uses: mamba-org/setup-micromamba@v1
         with:
           micromamba-version: '1.3.1-0'
-          environment-name: foo 
+          environment-file: environment.yml
           init-shell: >-
             bash
             powershell
@@ -27,8 +27,6 @@ jobs:
           post-cleanup: 'all'
       - name: Run python testing
         run: 
-          mamba activate foo
-          mamba install cdo nco
           pip install xarray
           pip install rioxarray
           pip install pytest
