+++ .github/workflows/python-app-mamba.yml
@@ -19,7 +19,7 @@ jobs:
       - uses: mamba-org/setup-micromamba@v1
         with:
           micromamba-version: '1.3.1-0'
-          environment-file: environment.yml
+          environment-file: .github/workflows/environment.yml
           init-shell: >-
             bash
             powershell
