### 2023-06-07 b97caea
use create-args to generate env in mamba test
