+++ .github/workflows/python-app-mamba.yml
@@ -19,7 +19,8 @@ jobs:
       - uses: mamba-org/setup-micromamba@v1
         with:
           micromamba-version: '1.3.1-0'
-          environment-file: .github/workflows/environment.yml
+          environment-name: foo
+          create-args: cdo nco  geos
           init-shell: >-
             bash
             powershell
