### 2023-06-07 cbf4319
use create-args to generate env in mamba test
