+++ .github/workflows/python-app-mamba.yml
@@ -20,7 +20,7 @@ jobs:
         with:
           micromamba-version: '1.3.1-0'
           environment-name: foo
-          create-args: cdo nco  geos
+          create-args: cdo nco geos
           init-shell: >-
             bash
             powershell
