### 2023-06-07 6e90cbe
try bash not micromamba shell for mamba testing
