### 2023-06-07 ec14350
try bash not micromamba shell for mamba testing
