### 2023-06-07 dc576fe
ensure foo activated in mamba test
