+++ .github/workflows/python-app-mamba.yml
@@ -28,6 +28,7 @@ jobs:
           post-cleanup: 'all'
       - name: Run python testing
         run: 
+          mamba activate foo
           pip install xarray
           pip install rioxarray
           pip install pytest
