### 2023-06-07 5a01c75
drop mamba activation
