+++ .github/workflows/python-app-mamba.yml
@@ -23,7 +23,6 @@ jobs:
           create-args: cdo nco geos
           init-shell: >-
             bash
-            powershell
           cache-environment: true
           post-cleanup: 'all'
       - name: Run python testing
@@ -43,4 +42,4 @@ jobs:
           pip install coverage 
           pytest tests/test*.py
         #shell: micromamba-shell {0}
-        shell: bash -l {0} 
+        shell: micromamba-shell {0} 
