### 2023-06-07 ca00464
drop . install from mambat
