+++ .github/workflows/python-app-mamba.yml
@@ -34,6 +34,7 @@ jobs:
           pip install cartopy
           pip install dask
           pip install -r requirements.txt
+          pip install .
           pip install pandas
           pip install rioxarray
           pip install pytest
