+++ .github/workflows/python-app-mamba.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Linux testing 
+name: Linux testing using micromamba
 
 on:
   push:
@@ -17,6 +17,8 @@ jobs:
     steps:
 
       - uses: mamba-org/setup-micromamba@v1
+        env:
+          ACTIONS_ALLOW_UNSECURE_COMMANDS: true
         with:
           micromamba-version: '1.3.1-0'
           environment-name: foo
@@ -26,7 +28,8 @@ jobs:
           cache-environment: true
           post-cleanup: 'all'
       - name: Run python testing
-        run: 
+        shell: bash -l {0} 
+        run: |
           pip install xarray
           pip install rioxarray
           pip install pytest
@@ -42,4 +45,3 @@ jobs:
           pip install coverage 
           pytest tests/test*.py
         #shell: micromamba-shell {0}
-        shell: micromamba-shell {0} 
