### 2023-06-07 9a47797
try allowing softlinks in mamba testing
