+++ .github/workflows/python-app-mamba.yml
@@ -26,6 +26,7 @@ jobs:
           init-shell: >-
             bash
           cache-environment: true
+          allow-softlinks: true
           post-cleanup: 'all'
       - name: Run python testing
         shell: bash -l {0} 
