### 2023-06-07 7cb2725
change install paths in mamba
