### 2023-06-07 d662875
test: print dir contents in mmaba test
