+++ .github/workflows/python-app-mamba.yml
@@ -30,18 +30,19 @@ jobs:
       - name: Run python testing
         shell: bash -l {0} 
         run: |
-          pip install xarray
-          pip install rioxarray
-          pip install pytest
-          pip install numpy 
-          pip install cartopy
-          pip install dask
-          pip install -r requirements.txt
-          pip install ../.
-          pip install pandas
-          pip install rioxarray
-          pip install pytest
-          pip install pytest-cov
-          pip install coverage 
-          pytest ../tests/test*.py
+          ls *
+          #pip install xarray
+          #pip install rioxarray
+          #pip install pytest
+          #pip install numpy 
+          #pip install cartopy
+          #pip install dask
+          #pip install -r requirements.txt
+          #pip install ../.
+          #pip install pandas
+          #pip install rioxarray
+          #pip install pytest
+          #pip install pytest-cov
+          #pip install coverage 
+          #pytest ../tests/test*.py
         #shell: micromamba-shell {0}
