### 2023-06-07 bf28f19
try modify bash shell
