+++ .github/workflows/python-app-mamba.yml
@@ -28,21 +28,21 @@ jobs:
           cache-environment: true
           post-cleanup: 'all'
       - name: Run python testing
-        shell: bash -l {0} 
+        shell: bash -el {0} 
         run: |
-          ls *
-          #pip install xarray
-          #pip install rioxarray
-          #pip install pytest
-          #pip install numpy 
-          #pip install cartopy
-          #pip install dask
-          #pip install -r requirements.txt
-          #pip install ../.
-          #pip install pandas
-          #pip install rioxarray
-          #pip install pytest
-          #pip install pytest-cov
-          #pip install coverage 
-          #pytest ../tests/test*.py
+          #ls *
+          pip install xarray
+          pip install rioxarray
+          pip install pytest
+          pip install numpy 
+          pip install cartopy
+          pip install dask
+          pip install -r requirements.txt
+          pip install .
+          pip install pandas
+          pip install rioxarray
+          pip install pytest
+          pip install pytest-cov
+          pip install coverage 
+          pytest ../tests/test*.py
         #shell: micromamba-shell {0}
