### 2023-06-07 9a2c098
try powershell in mamba t
