+++ .github/workflows/python-app-mamba.yml
@@ -25,10 +25,12 @@ jobs:
           create-args: cdo nco geos
           init-shell: >-
             bash
+            powershell
           cache-environment: true
           post-cleanup: 'all'
       - name: Run python testing
-        shell: bash -el {0} 
+        #shell: bash -el {0} 
+        shell: pwsh 
         run: |
           #ls *
           pip install xarray
