+++ .github/workflows/python-app-mamba.yml
@@ -23,6 +23,8 @@ jobs:
       - uses: mamba-org/setup-micromamba@v1
         env:
           ACTIONS_ALLOW_UNSECURE_COMMANDS: true
+      - uses: actions/checkout@v3
+      - uses: ./
         with:
           micromamba-version: '1.3.1-0'
           environment-name: foo
@@ -32,25 +34,58 @@ jobs:
             powershell
           cache-environment: true
           post-cleanup: 'all'
-      - uses: ./
-      - name: Run python testing
-        #shell: bash -el {0} 
-        #shell: pwsh 
+      - name: Install dependencies
+        shell: bash -l {0}
         run: |
-          cd ..
-          ls *
-          pip install xarray
-          pip install rioxarray
-          pip install pytest
+          #conda install cdo pip nco geos -y
           pip install numpy 
           pip install cartopy
           pip install dask
-          pip install -r requirements.txt
+          #conda install numpy -y
+          #conda install cartopy -y
+          #conda install dask -y
+          #    conda install flake8
+          # if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
           pip install .
           pip install pandas
           pip install rioxarray
           pip install pytest
           pip install pytest-cov
           pip install coverage 
-          pytest ../tests/test*.py
-        #shell: micromamba-shell {0}
+
+          #- name: Lint with flake8
+          #  shell: bash -l {0}
+          #  run: |
+          #    conda activate foo
+          #    # stop the build if there are Python syntax errors or undefined names
+          #    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+          #    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+          #    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+      - name: Test with pytest
+        shell: bash -l {0}
+        run: |
+          # conda activate foo
+          #pytest tests/test*.py
+          # python testing/test_parallel.py
+          pytest tests/test*.py
+      #- uses: ./
+      #- name: Run python testing
+      #  shell: bash -l {0} 
+      #  #shell: pwsh 
+      #  run: |
+      #    ls *
+      #    pip install xarray
+      #    pip install rioxarray
+      #    pip install pytest
+      #    pip install numpy 
+      #    pip install cartopy
+      #    pip install dask
+      #    pip install -r requirements.txt
+      #    pip install .
+      #    pip install pandas
+      #    pip install rioxarray
+      #    pip install pytest
+      #    pip install pytest-cov
+      #    pip install coverage 
+      #    pytest ../tests/test*.py
+      #  #shell: micromamba-shell {0}
