### 2023-06-07 1cdb337
test: drop checkout v3
