### 2023-06-07 b3811f4
test: drop checkout v3
