+++ .github/workflows/python-app-mamba.yml
@@ -19,7 +19,6 @@ jobs:
     runs-on: "ubuntu-latest" 
 
     steps:
-
       - uses: mamba-org/setup-micromamba@v1
         env:
           ACTIONS_ALLOW_UNSECURE_COMMANDS: true
@@ -32,19 +31,12 @@ jobs:
             powershell
           cache-environment: true
           post-cleanup: 'all'
-       uses: actions/checkout@v3
       - name: Install dependencies
         shell: bash -l {0}
         run: |
-          #conda install cdo pip nco geos -y
           pip install numpy 
           pip install cartopy
           pip install dask
-          #conda install numpy -y
-          #conda install cartopy -y
-          #conda install dask -y
-          #    conda install flake8
-          # if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
           pip install .
           pip install pandas
           pip install rioxarray
@@ -52,39 +44,7 @@ jobs:
           pip install pytest-cov
           pip install coverage 
 
-          #- name: Lint with flake8
-          #  shell: bash -l {0}
-          #  run: |
-          #    conda activate foo
-          #    # stop the build if there are Python syntax errors or undefined names
-          #    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-          #    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          #    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         shell: bash -l {0}
         run: |
-          # conda activate foo
-          #pytest tests/test*.py
-          # python testing/test_parallel.py
           pytest tests/test*.py
-      #- uses: ./
-      #- name: Run python testing
-      #  shell: bash -l {0} 
-      #  #shell: pwsh 
-      #  run: |
-      #    ls *
-      #    pip install xarray
-      #    pip install rioxarray
-      #    pip install pytest
-      #    pip install numpy 
-      #    pip install cartopy
-      #    pip install dask
-      #    pip install -r requirements.txt
-      #    pip install .
-      #    pip install pandas
-      #    pip install rioxarray
-      #    pip install pytest
-      #    pip install pytest-cov
-      #    pip install coverage 
-      #    pytest ../tests/test*.py
-      #  #shell: micromamba-shell {0}
