### 2023-06-07 76170a4
test: python pip install in mamba
