### 2023-06-07 pmlmodelling/nctoolkit#70 @robertjwilson
Thanks for the suggestion @malmans2. Figuring out if tests can be switched to mamba. I've managed to get that working this afternoon, after much tweaking with the yml. It knocked about 15 minutes of the run time: https://github.com/pmlmodelling/nctoolkit/actions/runs/5202108978.

Overall the tests are too thorough, so I can probably get the overall run time below 10 minutes easily enough, now that installs aren't an issue.
