### 2023-06-07 1da07d8
add macos mamba testing
