+++ .github/workflows/python-app-mamba-macos.yml
@@ -1,0 +1,53 @@
+# This workflow will install Python dependencies, run tests and lint with a single version of Python
+# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
+
+name: macOS testing using micromamba
+
+defaults:
+  run:
+    shell: bash -l {0}
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+
+    runs-on: " "macos-latest" " 
+
+    steps:
+      - uses: actions/checkout@v3
+      - uses: mamba-org/setup-micromamba@v1
+      # - uses: mamba-org/provision-with-micromamba@v15
+        env:
+          ACTIONS_ALLOW_UNSECURE_COMMANDS: true
+        with:
+          micromamba-version: '1.3.1-0'
+          environment-name: foo
+          create-args: cdo nco geos
+          init-shell: >-
+            bash
+            powershell
+          cache-environment: true
+          post-cleanup: 'all'
+      - name: Install dependencies
+        shell: bash -l {0}
+        run: |
+          pip install numpy 
+          pip install cartopy
+          pip install dask
+          python -m pip install .
+          # pip install .
+          pip install pandas
+          pip install rioxarray
+          pip install pytest
+          pip install pytest-cov
+          pip install coverage 
+
+      - name: Test with pytest
+        shell: bash -l {0}
+        run: |
+          pytest tests/test*.py
