### 2023-06-07 3e350ea
add macos mamba testing
