+++ .github/workflows/python-app-mamba-macos.yml
@@ -15,8 +15,7 @@ on:
 
 jobs:
   build:
-
-    runs-on: " "macos-latest" " 
+    runs-on:  "macos-latest"  
 
     steps:
       - uses: actions/checkout@v3
