### 2023-06-07 pmlmodelling/nctoolkit#70 @malmans2
Nice!
This software could be useful to debug GH actions: https://github.com/nektos/act
I've never used it myself though, I just thought you might find it useful.
If you end up using it, let me know how it goes!
