### 2023-06-08 pmlmodelling/nctoolkit#72 @malmans2
[JOSS] Add version to the cheat sheet
The cheat sheet is very helpful!
Is the cheatsheet in sync with the code?
Ideally, you need a GH action that generates it.
But if that's not feasible, I suggest to mention `nctoolkit` version in the cheat sheet.

https://github.com/openjournals/joss-reviews/issues/5494
