### 2023-06-08 pmlmodelling/nctoolkit#72 @robertjwilson
It's in synch, but agreed, it needs a version number.

I've been thinking of rewriting it in html, so that it can go on the website. Once I do that it might be feasible to autogenerate, or at least a test on GH actions to ensure it's up to date.
