### 2023-06-08 pmlmodelling/nctoolkit#73 @malmans2
[JOSS] Add ds.bottom?
I was looking at the cheatsheet, and I noticed `ds.top()`.
I'd find `ds.bottom()` very useful as it's a diagnostic more complicated to get for ocean data because the index changes.

This is more of a feature request than a review comment.

https://github.com/openjournals/joss-reviews/issues/5494
