### 2023-06-08 pmlmodelling/nctoolkit#74 @malmans2
[JOSS] Why is Windows not supported?
Why is Windows not supported? Is it a nightmare to maintain?
I just want to make sure that you considered supporting it, I understand that it might be high effort/low reward.
So feel free to close if there's no point in supporting windows.

https://github.com/openjournals/joss-reviews/issues/5494
