### 2023-06-08 #5494 @malmans2
Hi @robertjwilson,

I have this item in my checklist:

> Does the full list of paper authors seem appropriate and complete?

Could you please clarify very briefly the contribution of your co-author?
I'm asking because GH only shows a minor contribution: https://github.com/pmlmodelling/nctoolkit/commit/6ca02c98ea8ea1edf90beed294e234453d62a297

Thanks!
