### 2023-06-08 pmlmodelling/nctoolkit#68 @malmans2
Another comment about the example shown in the paper, looks like you already have a method to read/download from an url.
So I'd suggest to use that in the example rather than using a local path to a pre-downloaded file.
