### 2023-06-08 pmlmodelling/nctoolkit#68 @platipodium
related to #58
