### 2023-06-08 pmlmodelling/nctoolkit#69 @platipodium
I support this separation of concerns.
