+++ tests/test_merids.py
@@ -34,7 +34,7 @@ class TestClip:
         data.meridonial_mean()
         data.spatial_mean()
         if cdo_version() not in ["1.9.3", "1.9.4"]:
-            assert data.to_dataframe().sst[0].astype("float") == 17.67996597290039
+            assert data.to_dataframe().sst[0].astype("float") == 15.605058670043945 
         else:
             assert data.to_dataframe().sst[0].astype("float") == 17.679880142211914
 
@@ -44,7 +44,7 @@ class TestClip:
         data.meridonial_min()
         data.spatial_mean()
         if cdo_version() not in ["1.9.3", "1.9.4"]:
-            assert (data.to_dataframe().sst[0].astype("float") == -1.771429419517517)
+            assert (data.to_dataframe().sst[0].astype("float") == 6.8090081214904785)
         else:
             assert (data.to_dataframe().sst[0].astype("float") == -1.7714282274246216)
 
@@ -54,7 +54,7 @@ class TestClip:
         data.meridonial_max()
         data.spatial_mean()
         if cdo_version() not in ["1.9.3", "1.9.4"]:
-            assert data.to_dataframe().sst[0].astype("float") == 27.99681282043457
+            assert data.to_dataframe().sst[0].astype("float") == 21.77419662475586  
         else:
             assert data.to_dataframe().sst[0].astype("float") == 27.9967041015625
 
@@ -64,7 +64,7 @@ class TestClip:
         data.meridonial_range()
         data.spatial_mean()
         if cdo_version() not in ["1.9.3", "1.9.4"]:
-            assert data.to_dataframe().sst[0].astype("float") == 29.76824188232422
+            assert data.to_dataframe().sst[0].astype("float") == 14.965188026428223 
         else:
             assert data.to_dataframe().sst[0].astype("float") == 29.76813316345215
         del data
