### 2023-06-08 5c5a129
improve pubplot docstring
