### 2023-06-08 f195687
install cartopy using conda in all installs
