+++ .github/workflows/python-app-mamba-macos.yml
@@ -26,7 +26,7 @@ jobs:
         with:
           micromamba-version: '1.3.1-0'
           environment-name: foo
-          create-args: cdo nco geos
+          create-args: cdo nco cartopy
           init-shell: >-
             bash
             powershell
@@ -36,7 +36,6 @@ jobs:
         shell: bash -l {0}
         run: |
           pip install numpy 
-          pip install cartopy
           pip install dask
           python -m pip install .
           # pip install .
