+++ .github/workflows/python-app-mamba.yml
@@ -27,7 +27,7 @@ jobs:
         with:
           micromamba-version: '1.3.1-0'
           environment-name: foo
-          create-args: cdo nco geos
+          create-args: cdo nco cartopy
           init-shell: >-
             bash
             powershell
@@ -37,7 +37,6 @@ jobs:
         shell: bash -l {0}
         run: |
           pip install numpy 
-          pip install cartopy
           pip install dask
           python -m pip install .
           # pip install .
