### 2023-06-08 d2cdf12
drop parallel nco test
