### 2023-06-08 935095d
make pub_plot test work with different matplotlib v
