### 2023-06-08 cdaf299
drop similarity check for pubplot
