### 2023-06-08 f9ca531
fully replace linux testing with mamba
