+++ .github/workflows/python-app-linux.yml
@@ -1,79 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: Linux testing 
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-
-    runs-on: "ubuntu-latest" 
-
-    steps:
-
-     - uses: goanpeca/setup-miniconda@v1
-       env:
-        ACTIONS_ALLOW_UNSECURE_COMMANDS: true
-       with:
-         activate-environment: foo
-         python-version: 3.8
-         channels: conda-forge
-         allow-softlinks: true
-         channel-priority: flexible
-         show-channel-urls: true
-         use-only-tar-bz2: true
-
-     - uses: actions/checkout@v2
-       with:
-         activate-conda: true
-       #- name: Set up Python 3.8
-       #  uses: actions/setup-python@v2
-       #  with:
-       #    python-version: 3.8
-     - name: Install dependencies
-       shell: bash -l {0}
-       run: |
-         conda activate foo
-         conda install cdo pip nco cartopy -y
-         pip install numpy 
-         pip install dask
-         #conda install numpy -y
-         #conda install cartopy -y
-         #conda install dask -y
-         #    conda install flake8
-         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-         pip install .
-         pip install pandas
-         pip install rioxarray
-         pip install pytest
-         pip install pytest-cov
-         pip install coverage 
-
-         #- name: Lint with flake8
-         #  shell: bash -l {0}
-         #  run: |
-         #    conda activate foo
-         #    # stop the build if there are Python syntax errors or undefined names
-         #    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-         #    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-         #    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-     - name: Test with pytest
-       shell: bash -l {0}
-       run: |
-         conda activate foo
-         #pytest tests/test*.py
-         python testing/test_parallel.py
-         coverage run -m pytest tests/test*.py
-         coverage report
-         coverage xml
-
-     - name: Upload Coverage to Codecov
-       uses: codecov/codecov-action@v2
-
-
-
