+++ .github/workflows/python-app-mamba.yml
@@ -49,7 +49,6 @@ jobs:
       - name: Test with pytest
         shell: bash -l {0}
         run: |
-          pytest tests/test*.py
           python testing/test_parallel.py
           coverage run -m pytest tests/test*.py
           coverage report
