+++ .github/workflows/python-app-mamba.yml
@@ -19,7 +19,7 @@ jobs:
     runs-on: "ubuntu-latest" 
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v2
       - uses: mamba-org/setup-micromamba@v1
       # - uses: mamba-org/provision-with-micromamba@v15
         env:
