### 2023-06-09 07b473d
remove mac os testing -replaced by mamba
