+++ .github/workflows/python-app.yml
@@ -1,71 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: macOS testing 
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-
-    runs-on: "macos-latest" 
-
-    steps:
-
-     - uses: goanpeca/setup-miniconda@v1
-       env:
-        ACTIONS_ALLOW_UNSECURE_COMMANDS: true
-       with:
-         python-version: 3.11
-         activate-environment: foo
-         channels: conda-forge
-         allow-softlinks: true
-         channel-priority: flexible
-         show-channel-urls: true
-         use-only-tar-bz2: true
-
-     - uses: actions/checkout@v2
-       with:
-         activate-conda: true
-           #- name: Set up Python 3.8
-           #   uses: actions/setup-python@v2
-           #   with:
-           #     python-version: 3.10
-     - name: Install dependencies
-       shell: bash -l {0}
-       run: |
-         conda create -n foo python=3.10 nco cdo -y 
-         conda activate foo
-         #conda install cdo nco
-         #conda install pip
-         pip install pytest
-         #conda install numpy=1.23.3 -y
-         pip install dask 
-         #    conda install flake8
-         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-         #pip install numpy==1.23.3
-         pip install .
-         #pip install pandas==1.3.3
-         pip install rioxarray
-         pip install pytest-cov
-         conda install cartopy -y
-
-         #- name: Lint with flake8
-         #  shell: bash -l {0}
-         #  run: |
-         #    conda activate foo
-         #    # stop the build if there are Python syntax errors or undefined names
-         #    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-         #    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-         #    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-     - name: Test with pytest
-       shell: bash -l {0}
-       run: |
-         conda activate foo
-         python testing/test_scripting.py
-         python testing/test_parallel.py
-         pytest tests/test_*.py
