### 2023-06-09 pmlmodelling/nctoolkit#72 @robertjwilson
The version number is [now added](https://github.com/pmlmodelling/nctoolkit/blob/master/cheatsheet/nctoolkit_cheatsheet.pdf)

I have raised an issue to generate the cheatsheets using a similar approach to RStudio, who use Quarto which will pull in version numbers etc. https://github.com/pmlmodelling/nctoolkit/issues/75
