### 2023-06-09 pmlmodelling/nctoolkit#74 @robertjwilson
CDO does not have native Windows support: https://code.mpimet.mpg.de/projects/cdo/wiki/Win32. It can be installed using cygwin on Windows, but supporting this in nctoolkit would be pretty difficult/elaborate and not likely worth it as netCDF users primarily work on Linux.
