### 2023-06-09 pmlmodelling/nctoolkit#73 @robertjwilson
A `ds.bottom` method already exists: https://nctoolkit.readthedocs.io/en/latest/generated/nctoolkit.DataSet.bottom.html?highlight=bottom#nctoolkit.DataSet.bottom

That method will only extract the actual bottom level in the dataset.

For ocean data, there is a separate method for creating a mask which will identify the bottom cell in ocean data: https://nctoolkit.readthedocs.io/en/latest/generated/nctoolkit.DataSet.bottom_mask.html?highlight=bottom_mask

An extra arg should possibly be added to `bottom`, so that it can call when `bottom_mask` when you are dealing with oceanic data etc.
