+++ .github/workflows/python-app-macos.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: macOS testing using micromamba
+name: macOS tests
 
 defaults:
   run:
