+++ .github/workflows/python-app-linux.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Linux testing using micromamba
+name: Linux tests
 
 defaults:
   run:
