### 2023-06-09 pmlmodelling/nctoolkit#73 @malmans2
Got it, thanks.
If your code (or CDO) don't look for that variable already, keep in mind that some NetCDF could have a variable that defines the bottom. The CF standard name is `model_level_number_at_sea_floor`
