### 2023-06-09 c7799d1
add furo as requirement to try it
