+++ docs/requirements.txt
@@ -14,6 +14,7 @@ ncplot
 numpy
 plotnine
 dill
+furo
 
 
 
