+++ docs/source/conf.py
@@ -39,7 +39,8 @@ extensions = [
 ]
 
 #html_theme = "sphinx_rtd_theme"
-html_theme = "furo"
+#html_theme = "furo"
+html_theme = "press"
 html_static_path = ['_static']
 html_logo = "_static/nctoolkit_logo.jpg"
 
