+++ docs/source/conf.py
@@ -40,8 +40,9 @@ extensions = [
 
 #html_theme = "sphinx_rtd_theme"
 #html_theme = "furo"
-html_permalinks_icon = '<span>#</span>'
-html_theme = 'sphinxawesome_theme'
+import sphinx_bootstrap_theme
+html_theme = 'bootstrap'
+html_theme_path = sphinx_bootstrap_theme.get_html_theme_path()
 #html_theme = "better"
 html_static_path = ['_static']
 html_logo = "_static/nctoolkit_logo.jpg"
