+++ docs/requirements.txt
@@ -14,7 +14,7 @@ ncplot
 numpy
 plotnine
 dill
-sphinx-bootstrap-theme
+sphinx-theme-pd
 #furo
 #sphinx-book-theme
 sphinx-better-theme
