+++ docs/source/conf.py
@@ -40,10 +40,10 @@ extensions = [
 
 #html_theme = "sphinx_rtd_theme"
 #html_theme = "furo"
-import sphinx_theme_pd
-html_theme = 'sphinx_theme_pd'
-html_theme_path = [sphinx_theme_pd.get_html_theme_path()]
-#html_theme = "better"
+import sphinx_pdj_theme
+html_theme = 'sphinx_pdj_theme'
+html_theme_path = [sphinx_pdj_theme.get_html_theme_path()]
+
 html_static_path = ['_static']
 html_logo = "_static/nctoolkit_logo.jpg"
 
