+++ docs/source/conf.py
@@ -40,9 +40,7 @@ extensions = [
 
 #html_theme = "sphinx_rtd_theme"
 #html_theme = "furo"
-import sphinx_pdj_theme
-html_theme = 'sphinx_pdj_theme'
-html_theme_path = [sphinx_pdj_theme.get_html_theme_path()]
+html_theme = 'haiku'
 
 html_static_path = ['_static']
 html_logo = "_static/nctoolkit_logo.jpg"
