+++ docs/requirements.txt
@@ -1,10 +1,13 @@
-sphinx==4.0.2
+#sphinx==4.0.2
+sphinx
 numpydoc==0.9.2
-ipykernel==5.1.2
+#ipykernel==5.1.2
+ipykernel
 nbsphinx>=0.8.6
 Pygments==2.7.4
 sphinx_rtd_theme==0.5.2
-docutils<0.17
+#docutils<0.17
+docutils
 widgetsnbextension
 ipywidgets
 xarray
