+++ docs/requirements.txt
@@ -1,5 +1,5 @@
-#sphinx==4.0.2
-sphinx
+sphinx==4.0.2
+#sphinx
 numpydoc==0.9.2
 #ipykernel==5.1.2
 ipykernel
