### 2023-06-09 b085e19
add plots to interpolations rst
