+++ docs/source/interpolation1.rst
@@ -21,11 +21,10 @@ The data is available using a thredds server. So we will work with the first tim
    ds.plot()
 
 .. raw:: html
-   :file: visualization_plot1.html
+   :file: interpolation_plot1.html 
 
 
 Interpolating to a set of coordinates
-
 --------------------------------------
 
 
@@ -92,14 +91,125 @@ with longitude -30 and latitude 50.
     </div>
 
 
+Interpolating to a regular lonlat grid
+---------------------------------------
 
-Plotting internals
+If you want to interpolate to a regular latlon grid, you can use ``to_latlon``. ``lon`` and ``lat`` specify the minimum and maximum longitudes and latitudes, while ``res``, a 3 variable list specifies the resolution. For example, if we wanted to regrid the globe to 0.5 degree north-south by 1 degree east-west resolution, we could do the following:
+
+
+.. code:: ipython3
+
+
+   ds = nc.open_thredds("https://psl.noaa.gov/thredds/dodsC/Datasets/COBE2/sst.mon.mean.nc")
+   ds.subset(timestep = 0)
+   ds.to_latlon(lon = [-79.5, 79.5], lat = [0.75, 89.75], res = [1, 0.5])
+   ds.plot()
+
+.. raw:: html
+   :file: interpolation_plot2.html
+
+
+Interpolating to another dataset’s grid
+---------------------------------------
+If we are working with two datasets and want to put them on a common grid, we can interpolate one onto the other’s grid. We can illustate this with a dataset of global sea surface temperature. Let’s start by cropping this dataset to the northern hemisphere. 
+
+
+.. code:: ipython3
+   ds1 = nc.open_thredds("https://psl.noaa.gov/thredds/dodsC/Datasets/COBE2/sst.mon.mean.nc")
+   ds1.subset(timestep = 0)
+   ds1.subset(lat = [0, 90]) 
+   ds1.plot()
+
+.. raw:: html
+   :file: interpolation_plot3.html
+
+
+Now, we can regrid the original file to this northern hemisphere grid.
+
+.. code:: ipython3
+   ds2 = nc.open_thredds("https://psl.noaa.gov/thredds/dodsC/Datasets/COBE2/sst.mon.mean.nc")
+   ds2.subset(timestep = 0)
+   ds2.regrid(ds1)
+   ds2.plot()
+
+
+.. raw:: html
+   :file: interpolation_plot4.html
+
+This method will also work using netCDF files. So, if you wanted you can also use a path to a netCDF file as the target grid.
+
+
+How to reuse the weights for regridding
+---------------------------------------
+Under the hood nctoolkit regrids data by first generating a weights file. There are situations where you  will want to be able to re-use these weights. For example, if you are post-processing a large number of files one after the other. To make this easier nctoolkit let's you recycle the regridding info. This let's you interpolate using either ``regrid`` or ``to_latlon``, but keep the regridding data for future use by ``regrid``.
+
+
+
+.. code:: ipython3
+
+   ds = nc.open_thredds("https://psl.noaa.gov/thredds/dodsC/Datasets/COBE2/sst.mon.mean.nc")
+   ds.subset(timestep = 0)
+   ds.to_latlon(lon = [-79.5, 79.5], lat = [-0.75, 89.75], res = [1, 0.5], recycle = True)
+   ds.plot()
+.. raw:: html
+   :file: interpolation_plot5.html
+
+
+.. code:: ipython3
+   ds1 = nc.open_thredds("https://psl.noaa.gov/thredds/dodsC/Datasets/COBE2/sst.mon.mean.nc")
+   ds1.subset(timestep = 0)
+   ds1.regrid(ds)
+   ds1.plot()
+
+.. raw:: html
+   :file: interpolation_plot6.html
+
+Horizontal Resampling
 ---------------------
[…]
