### 2023-06-09 4740f8d
fix ipython error in lines
