### 2023-06-09 f0bfd89
ipython tweek for interpo
