+++ docs/requirements.txt
@@ -19,3 +19,4 @@ plotnine
 dill
 sphinx-book-theme
 
+setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
