### 2023-06-12 8be4657
add ability to pass fig to pub_plot
