### 2023-06-12 4cdad0b
add break arg to pub_plot
