### 2023-06-12 d263a30
qdd quiver plot
