+++ nctoolkit/static_plot.py
@@ -201,6 +201,17 @@ def pub_plot(
         gs = None
 
 
+    if "quiver" in kwargs:
+        quiver = True
+        u = kwargs["u"]
+        v = kwargs["v"]
+        kwargs.pop("quiver")
+        kwargs.pop("u")
+        kwargs.pop("v")
+    else:
+        quiver = False
+
+
     for kk in kwargs:
         fixed = False
 
@@ -274,14 +285,15 @@ def pub_plot(
 
     ds1 = ds.copy()
 
-    if var is not None:
-        ds1.subset(variables=var)
-    ds1.run()
+    if not quiver:
+        if var is not None:
+            ds1.subset(variables=var)
+        ds1.run()
 
-    if len(ds1.variables) > 1:
-        raise ValueError("Only one variable wanted")
-    if len(ds1.times) > 1:
-        raise ValueError("Only one timestep wanted")
+        if len(ds1.variables) > 1:
+            raise ValueError("Only one variable wanted")
+        if len(ds1.times) > 1:
+            raise ValueError("Only one timestep wanted")
 
     ds_xr = ds1.to_xarray(decode_times=False)
     failed = False
@@ -402,54 +414,64 @@ def pub_plot(
         pass
 
     # read the values. the array is squeezed to remove the temporal dimension that has length 1
-    values = input_file.variables[ds1.variables[0]][:].squeeze()
+    if quiver:
+        u = input_file.variables[u][:].squeeze()
+        v = input_file.variables[v][:].squeeze()
 
-    # mask where the values is 0 == land points
-    values = np.ma.masked_where(values == 0, values)
+        # mask where the values is 0 == land points
+        # values = np.ma.masked_where(values == 0, values)
+        u = np.ma.masked_where(u == 0, u)
+        v = np.ma.masked_where(v == 0, v)
 
-    if limits != None:
-        if limits[0] is None:
-            limits[0] = np.min(values)
-        if limits[1] is None:
-            limits[1] = np.max(values)
+    if not quiver:
+        values = input_file.variables[ds1.variables[0]][:].squeeze()
 
-    if grid_colour == "auto":
-        value_10 = values.min() + (values.max() - values.min()) * 0.3
-        if np.mean(values.data < value_10) < 0.7:
-            grid_colour = "black"
+        # mask where the values is 0 == land points
+        values = np.ma.masked_where(values == 0, values)
 
-    r_min = None
-    r_max = None
+        if limits != None:
+            if limits[0] is None:
+                limits[0] = np.min(values)
+            if limits[1] is None:
+                limits[1] = np.max(values)
 
-    if limits is not None:
-        if isinstance(limits[0], str):
-            if "%" in limits[0]:
-                r_min = float(limits[0].split("%")[0])
-        if isinstance(limits[1], str):
-            if "%" in limits[1]:
-                r_max = float(limits[1].split("%")[0])
+        if grid_colour == "auto":
+            value_10 = values.min() + (values.max() - values.min()) * 0.3
+            if np.mean(values.data < value_10) < 0.7:
+                grid_colour = "black"
 
-    if mid_point is None and limits is None:
-        min_value = np.min(values)
-        max_value = np.max(values)
-        if robust:
-            min_value = np.nanpercentile(np.ma.filled(values, np.nan), 2)
-            max_value = np.nanpercentile(np.ma.filled(values, np.nan), 98)
-        if r_min is not None:
-            min_value = np.nanpercentile(np.ma.filled(values, np.nan), r_min)
-        if r_max is not None:
-            max_value = np.nanpercentile(np.ma.filled(values, np.nan), r_max)
+        r_min = None
+        r_max = None
 
-        if min_value < 0 and max_value > 0:
[…]
