### 2023-06-13 pmlmodelling/nctoolkit#77 @malmans2
[JOSS] Object representation does not reflect lazy operations
**Describe the bug**
Looks like I'm not able to select a single year from a NCEP dataset (or at least the representation of the DataSet object does not show the subsetting).

**To Reproduce**
```python
import nctoolkit as nc
nc_ds = nc.open_url("https://github.com/pydata/xarray-data/raw/master/air_temperature.nc")
print(nc_ds)
nc_ds.subset(year = 2013)
print(nc_ds)
```

    nctoolkit is using Climate Data Operators version 2.2.0
    Downloading https://github.com/pydata/xarray-data/raw/master/air_temperature.nc

    The variable air has integer data type. Consider setting data type to float 'F64' or 'F32' using set_precision.

    <nctoolkit.DataSet>:
    Number of files: 1
    File contents:
      variable  ntimes  npoints  nlevels                                   long_name  unit data_type
    0      air    2920     1325        1  4xDaily Air temperature at sigma level 995  degK       I16

    <nctoolkit.DataSet>:
    Number of files: 1
    File contents:
      variable  ntimes  npoints  nlevels                                   long_name  unit data_type
    0      air    2920     1325        1  4xDaily Air temperature at sigma level 995  degK       I16

**Expected behavior**
`ntimes` should change from 2920 to 1460
```python
import xarray as xr
xr_ds = xr.tutorial.open_dataset("air_temperature").chunk()
print(xr_ds.dims)
xr_ds = xr_ds.sel(time="2013")
print(xr_ds.dims)
```

    Frozen({'lat': 25, 'time': 2920, 'lon': 53})
    Frozen({'lat': 25, 'time': 1460, 'lon': 53})

 - OS: macOS
 - nctoolkit version: '0.9.3'

https://github.com/openjournals/joss-reviews/issues/5494
