### 2023-06-13 pmlmodelling/nctoolkit#78 @malmans2
[JOSS] Documentation notebooks
Are the documentation notebooks executed by readthedocs?
If not, I suggest to do that and to store notebooks without any output (e.g., I often use [nbstripout with pre-commit](https://github.com/kynan/nbstripout#using-nbstripout-as-a-pre-commit-hook)).

If you need to show notebooks that need a long run time, I suggest to always print `nctoolkit` version at import.

https://github.com/openjournals/joss-reviews/issues/5494
