### 2023-06-13 pmlmodelling/nctoolkit#77 @malmans2
I see, I need to run `nc_ds.run()` to actually see the changes.
I find this quite confusing, especially because dataset objects are modified in place and therefore it's very hard to keep track of the modifications that will be applied.

I think that the representation of the object should show the modified coordinates/dimensions (same as `xarray`+`dask`, which is also lazy), or it should show at least all the operations that will be applied when `nc_ds.run()` is called.
