### 2023-06-19 a066f70
add type checks to quiver plot
