### 2023-06-19 419c1c6
add ability to use fig without gs in pub_plot
