### 2023-06-20 pmlmodelling/nctoolkit#73 @robertjwilson
Thanks. I'll put using that in `mask` onto my to-do list.
