### 2023-06-20 pmlmodelling/nctoolkit#57 @robertjwilson
The simple answer is that the package is currently developed based on project needs. Using `dask` is something on our minds, but it will need to align with ongoing project needs.
