### 2023-06-20 pmlmodelling/nctoolkit#69 @robertjwilson
I agree the dependencies cause some issues, though mostly with conda installations due to version resolving issues.

I don't however see how plotting is an optional dependency. This is something that 99% of users will have to use, so I'm not sure that splitting it out will do anything but annoy people.

I've spoken with a few users of the package to get views, and they seemed clear that separating this out would annoy them. This is especially true for people who work with fresh conda environments all the time, and they don't want to have to remember how to install it.

There is a case perhaps for a nctoolkit-core package on conda (much like geoviews-core etc.), which would side-step conda installation problems and just be for pure post-processing, not plotting. But it's clear to me if the time spent creating and then maintaing this are worth it, as the appeal is likely very limited.

Open to have my mind changed.
