### 2023-06-20 pmlmodelling/nctoolkit#78 @robertjwilson
The notebooks are currently run locally. Improving this is on my to-do list, but it's not a short piece of work so it's not high on the list. I'd probably need to rewrite a lot of them to use readthedocs as the files are a) downloaded and b) too large to store somewhere like Github.

I've closed the issue as it's out of scope for JOSS in my view, but I've raised this issue to ensure it's dealt with in future. https://github.com/pmlmodelling/nctoolkit/issues/79
