### 2023-06-20 pmlmodelling/nctoolkit#62 @robertjwilson
Can you clarify if this refers to the paper, website or Github @platipodium?
