### 2023-06-20 pmlmodelling/nctoolkit#65 @robertjwilson
Path is now fixed along with line break.
