### 2023-06-20 pmlmodelling/nctoolkit#80 @robertjwilson
Hi @malmans2. I am able to reproduce this. I've pinned it down to an upgrade to numpy 3 days ago breaking something within hvplot. Downgrading numpy to 1.24.3 seems to fix the problem.

I'll put together a reproducible example and raise an issue over at hvplot. It seems to only apply to cases with multiple depths, and the numpy change appears to be changing the behaviour of xarray which is causing the problem in hvplot.
