### 2023-06-20 pmlmodelling/nctoolkit#63 @platipodium
I think most users would not expect this information in the JOSS paper nor on the website (but possibly on readthedocs), clearly this should be resolved on the GitHub repo.

I would search for such information  at
- the end of the ReadMe
- in a file called `Reporting.md`,  `Bugs.md`, `Support.md`, `Contributing.md` or the like in the main directory (best referenced to from `ReadMe`

But JOSS requirement is only that such information is available (and, I guess, findable)
