### 2023-06-20 9a0d535
add plotting dependencies
