### 2023-06-20 pmlmodelling/nctoolkit#57 @platipodium
Could you then make a performance statement, or a statement on the limitations of your software and its multiprocessing capabilities?
