### 2023-06-20 99dd9b6
add scipy as noted dep
