### 2023-06-21 pmlmodelling/nctoolkit#81 @malmans2
[JOSS] Performance: Show that nctoolkit is fast
**Fast** and **easy** analysis of netCDF data in Python.

The fact that `nctoolkit` is **easy** to use is well shown and explained throughout the documentation.
Is there a section that explains and shows that it's also **fast**?
