### 2023-06-21 pmlmodelling/nctoolkit#84 @malmans2
Remove data from this repo
I noticed that cloning this repo is quite slow. I imagine that's because of the `data`.
Did you consider using [git-lfs](https://git-lfs.com/) or moving data to another repo?
