+++ .github/workflows/python-app-macos.yml
@@ -48,4 +48,5 @@ jobs:
       - name: Test with pytest
         shell: bash -l {0}
         run: |
+          python testing/test_parallel.py
           pytest tests/test*.py
