### 2023-06-21 6b8ecd6
add note on multiprocess for macos
