### 2023-06-21 pmlmodelling/nctoolkit#82 @robertjwilson
Thanks for spotting this. I changed the nctoolkit internals for macOS 2 weeks ago to fix this bug on macOS: https://github.com/pmlmodelling/nctoolkit/issues/29

But I forgot to update this doc.

Essentially, nctoolkit now uses multiprocess when running in parallel on macOS, but multiprocessing on Linux, due to weird differences in behaviour introduced in Python 3.7. So the example needs to use multiprocess on macOS. I've just added this to the docs: https://nctoolkit.readthedocs.io/en/latest/parallel.html
