### 2023-06-21 pmlmodelling/nctoolkit#82 @malmans2
OK, multiprocess does not crash like before.

But I think I'm misunderstanding this example.

This is copied and pasted from the documentation.
```python
import nctoolkit as nc
nc.options(parallel = True)

def process_chain(infile, outfile):
    ds = nc.open_data(ff)
    ds.assign(tos = lambda x: x.sst + 273.15)
    ds.tmean()
    ds.to_nc(outfile)

ensemble = nc.create_ensemble("data/ensemble")
import multiprocess as mp
pool = mp.Pool(2)
for ff in ensemble:
    pool.apply_async(process_chain, [ff, ff.replace("ensemble", "new")])
pool.close()
pool.join()
```

Is this assertion correct? Because it fails on my machine, so `data/new` does not exits.
```python
import os
assert os.path.exists("data/new")
```

Also, is this correct? `ds = nc.open_data(ff)`, or should it be `ds = nc.open_data(infile)`?
I tried both of them, but nothing fails and I don't see the new directory.
