### 2023-06-21 pmlmodelling/nctoolkit#82 @malmans2
Ahhh I see, I need to create the directory `new` myself, then it gets populated.
It worth mentioning it in the documentation, and I guess you need to fix the bug in the function (`ff` vs `infile`).

But I think there's still a problem because there are silent failures.
