### 2023-06-22 pmlmodelling/nctoolkit#70 @malmans2
I'll close this as the tests now take less than 10min.
Note that you still have a warning:

> Node.js 12 actions are deprecated. Please update the following actions to use Node.js 16: actions/checkout@v2, codecov/codecov-action@v2. For more information see: https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/.

You are not using the latest codecov version (you can use [dependabot](https://docs.github.com/en/code-security/dependabot/working-with-dependabot/keeping-your-actions-up-to-date-with-dependabot) to keep actions up to date).
