### 2023-06-22 pmlmodelling/nctoolkit#70 @robertjwilson
Thanks @malmans2. macOS tests are still running slow, because CDO is slower on macOS at least on GitHub's architecture. But I should maybe just switch those tests to daily builds.

I've raised an issue to get the dependabot issue fixed: https://github.com/pmlmodelling/nctoolkit/issues/85
