### 2023-06-22 #5494 @platipodium
I am d'accord with @malmans2, in principle there are no major obstacles for the publication.  There are still some open issues to be sorted out, and I should be able to tick them off my checklist (aligning pretty much with @malmans2 checklist) once they are addressed.

## Remaining items from checklist for @platipodium

- [ ] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?
