### 2023-06-22 f0e98de
speed up file opening in tests
