+++ tests/test_centre.py
@@ -12,7 +12,7 @@ ff = "data/sst.mon.mean.nc"
 
 class TestClip:
     def test_centre(self):
-        tracker = nc.open_data(ff)
+        tracker = nc.open_data(ff, checks = False)
         with pytest.raises(ValueError):
             tracker.centre("test")
         n = len(nc.session_files())
@@ -25,13 +25,13 @@ class TestClip:
             tracker = nc.open_data("data/ensemble/*.nc")
             tracker.centre(by_area = True)
 
-        tracker = nc.open_data(ff)
+        tracker = nc.open_data(ff, checks = False)
         tracker.centre("longitude")
         x = tracker.to_dataframe().sst.values[0].astype("float")
 
         assert np.round(x,4) == np.round(-31.178354263305664, 4)
 
-        tracker = nc.open_data(ff)
+        tracker = nc.open_data(ff, checks = False)
         tracker.centre("latitude")
         x = tracker.to_dataframe().sst.values[0].astype("float")
 
@@ -42,7 +42,7 @@ class TestClip:
 
 
 
-        tracker = nc.open_data(ff)
+        tracker = nc.open_data(ff, checks = False)
         tracker.assign(sst1 = lambda x: x.sst + 0)
         tracker.centre("latitude")
         x = tracker.to_dataframe().sst1.values[0].astype("float")
