### 2023-06-22 ac8d270
further improvements in file opening for tests
