### 2023-06-27 2e8f74a
replace fix_nemo_.. with amm7
