### 2023-06-27 2251d7d
ensure compatibility with numpy res
