### 2023-07-10 aa4f1c1
add ability to use depths in verticals
