+++ nctoolkit/tozlev.py
@@ -6,7 +6,7 @@ from nctoolkit.session import remove_safe
 import nctoolkit.api as api
 
 
-def to_zlevels(self, levels=None, thickness=None):
+def to_zlevels(self, levels=None, thickness=None, depths = None):
     """
     Convert datasets with non z-level verticals to z-levels
     Experimental method: Use at your own risk.
@@ -20,23 +20,30 @@ def to_zlevels(self, levels=None, thickness=None):
         One of: a variable, in the dataset, which contains the variable thicknesses; a .nc file which contains
         the thicknesses; or a Dataset that contains the thicknesses. Note: the .nc file or Dataset must only contain
         one variable. Thickness should be in metres. Vertical interpolation will take the value from the mid-point of the level.
+    depths: str or Dataset
+        This or thickness must be supplied if fixed is False, otherwise vertical thickness/depth cannot be known.
+        Option argument when vertical levels vary in space.
+        One of: a variable, in the dataset, which contains the variable depths; a .nc file which contains
+        the depths; or a Dataset that contains the depths. Note: the .nc file or Dataset must only contain
+        one variable. Depths should be in metres, and be the mid-point of the level.
     """
     self.run()
 
     if len(self) > 1:
         raise ValueError("This currently only works on single file datasets")
 
-    if thickness is None:
+    if thickness is None and depths is None:
         if "e3t" in self.variables:
             thickness = "e3t"
 
     drop_this = None
 
-    if not isinstance(thickness, api.DataSet):
-        if thickness is None or not isinstance(thickness, str):
-            raise ValueError("Please provide a valid thickness or depths variable")
+    if depths is None:
+        if not isinstance(thickness, api.DataSet):
+            if thickness is None or not isinstance(thickness, str):
+                raise ValueError("Please provide a valid thickness or depths variable")
 
-    if thickness is None:
+    if thickness is None and depths is None:
         if not isinstance(depths, api.DataSet):
             if depths is None or not isinstance(depths, str):
                 raise ValueError("Please provide a valid thickness or depths variable")
@@ -51,41 +58,51 @@ def to_zlevels(self, levels=None, thickness=None):
 
     sorted = False
 
-    if isinstance(thickness, api.DataSet):
-        ds_depths = thickness.copy()
-        ds_depths.run()
-        if len(ds_depths.variables) != 1:
-            raise ValueError("Please provide a thickness dataset with 1 variable!")
-        sorted = True
-
-    if sorted is False:
-        if thickness in self.variables:
-            ds_depths = open_data(self[0])
-            ds_depths.subset(variable=thickness)
+    if depths is None:
+        if isinstance(thickness, api.DataSet):
+            ds_depths = thickness.copy()
             ds_depths.run()
-            drop_this = thickness
-        else:
-            ds_depths = open_data(thickness)
             if len(ds_depths.variables) != 1:
-                raise ValueError("Please provide a thickness file with 1 variable!")
+                raise ValueError("Please provide a thickness dataset with 1 variable!")
+            sorted = True
+    else:
+        if isinstance(depths, api.DataSet):
+            ds_depths = depths.copy()
+            ds_depths.run()
+            if len(ds_depths.variables) != 1:
+                raise ValueError("Please provide a thickness or depths dataset with 1 variable!")
+            sorted = True
 
-    thick_var = ds_depths.variables[0]
+    if depths is None:
+        if sorted is False:
[…]
