### 2023-07-11 be47d19
switch from run_this to cdo_command in methods
