+++ nctoolkit/cdo_command.py
@@ -1,4 +1,5 @@
 import subprocess
+import os
 
 from nctoolkit.runthis import run_this
 
@@ -33,15 +34,29 @@ def cdo_command(self, command=None, ensemble=False):
 
     cdo_methods = [mm for mm in cdo_methods if len(mm) > 0]
 
+    cdo_methods.append("L")
+    cdo_methods.append("reduce_dim")
+
     for x in command.replace("-f ", "").replace("-z ", "").strip().split(" "):
         y = x.split(",")[0].replace("-", "")
         if y not in cdo_methods and y != "nc4" and y != "zip_9":
-            raise ValueError(f"{y} is not a cdo method!" + y)
+            if not os.path.exists(y):
+                raise ValueError(f"{y} is not a cdo method!" + y)
+
+    new_command = ""
+    for cc in command.split(" "):
+        if cc in cdo_methods:
+            new_command += " -" + cc
+        else:
+            new_command += " " + cc
+
+    new_command = new_command.strip()
+    new_command = new_command.replace("  ", " ")
 
     # remove cdo from the command
 
     output = "ensemble"
-    cdo_command = "cdo " + command + " "
+    cdo_command = "cdo " + new_command + " "
 
     if ("merge " in cdo_command) or ("mergetime " in cdo_command) or ensemble:
         output = "one"
@@ -50,5 +65,6 @@ def cdo_command(self, command=None, ensemble=False):
     for mm in cdo_methods:
         if " " + mm + "," in cdo_command:
             cdo_command = cdo_command.replace(" " + mm + ",", " -" + mm + ",")
+    cdo_command = cdo_command.strip()
 
     run_this(cdo_command, self, output=output)
