### 2023-07-12 414bd8c
add note about how to ask questions
