### 2023-07-14 c8477e0
update conda version instructions
