+++ docs/source/installing.rst
@@ -14,7 +14,7 @@ The best and easiest way to install nctoolkit is to use conda. This will install
 
 Conda can often install a very old version of nctoolkit. So you might want to install a recent version::
 
-   $ conda install -c conda-forge nctoolkit=0.9.0
+   $ conda install -c conda-forge nctoolkit=0.9.4
 
 Mamba is a smoother way to manage conda environments. If you don't use it, you should try. Install it from  `here  <https://github.com/conda-forge/miniforge/>`__.
 
