### 2023-07-18 pmlmodelling/nctoolkit#77 @robertjwilson
Yeah, there is some ambiguity here. I think the solution is to automatically run `ds.run()` when you access attributes etc. This behaviour would be more what a new user would expect. And it's not going to have any computational impacts, as you'll only really be accessing attributes interactively, not when scripting.

In theory, changes could be tracked without running commands, but that would just become very awkward book-keeping.
