### 2023-07-18 8b81163
run commands when attributes accessed
