+++ nctoolkit/api.py
@@ -119,6 +119,20 @@ session_info["progress"] = "on"
 session_info["checks"] = True
 session_info["user"] = ""
 
+# get the cdo methods
+
+read = subprocess.run(
+    "cdo --operators", shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+).stdout
+
+cdo_methods = [x.split(" ")[0].replace("b'", "") for x in str(read).split("\\n")]
+
+cdo_methods = [mm for mm in cdo_methods if len(mm) > 0]
+
+cdo_methods.append("L")
+cdo_methods.append("reduce_dim")
+session_info["cdo_methods"] = cdo_methods
+
 
 def coast_check():
     try:
@@ -331,87 +345,6 @@ def options(**kwargs):
     update_options(kwargs)
     return None
 
-    for key in kwargs:
-        if key not in valid_keys:
-            raise AttributeError(key + " is not a valid option")
-
-        if key == "parallel" or key == "lazy" or key == "thread_safe":
-            if isinstance(kwargs[key], bool):
-                raise TypeError(f"{key} should be boolean")
-
-        if key == "checks":
-            if isinstance(kwargs[key], bool):
-                raise TypeError(f"{key} should be boolean")
-
-        if isinstance(kwargs[key], bool):
-            if key == "temp_dir":
-                if isinstance(kwargs[key], str):
-                    if os.path.exists(kwargs[key]) is False:
-                        raise ValueError("The temp_dir specified does not exist!")
-                    session_info[key] = os.path.abspath(kwargs[key])
-                    session_info["user_dir"] = True
-                return None
-            if key == "progress":
-                if kwargs[key] not in ["on", "off", "auto", "of"]:
-                    raise ValueError("progress must be one of 'on', 'off', 'auto'")
-
-                if kwargs[key] == "of":
-                    session_info[key] = "off"
-                else:
-                    session_info[key] = kwargs[key]
-                return None
-
-            if key == "cores":
-                if isinstance(kwargs[key], int):
-                    if kwargs[key] > mp.cpu_count():
-                        raise ValueError(
-                            str(kwargs[key])
-                            + " is greater than the number of system cores ("
-                            + str(mp.cpu_count())
-                            + ")"
-                        )
-                    session_info[key] = kwargs[key]
-                else:
-                    raise TypeError("cores must be an int")
-            else:
-                if key == "precision":
-                    if kwargs[key] not in ["I8", "I16", "I32", "F32", "F64"]:
-                        raise ValueError("precision supplied is not valid!")
-                    session_info[key] = kwargs[key]
-                else:
-                    raise ValueError(kwargs[key] + " is not valid session info!")
-        else:
-            session_info[key] = kwargs[key]
-
-            # update safe-lists etc. if running in parallel
-            if kwargs[key] and key == "parallel":
-                if len(temp_dirs) > 0:
-                    for ff in temp_dirs:
-                        append_tempdirs(ff)
-
-                if len(nc_safe) > 0:
-                    for ff in nc_safe:
-                        nc_safe_par.append(ff)
-
-                if len(nc_protected) > 0:
-                    for ff in nc_protected:
-                        nc_protected_par.append(ff)
-                        nc_protected.remove(ff)
-
-            if (kwargs[key] is False) and key == "parallel":
-                if len(temp_dirs_par) > 0:
-                    for ff in temp_dirs_par:
-                        append_tempdirs(ff)
-
-                if len(nc_safe_par) > 0:
-                    for ff in nc_safe_par:
-                        nc_safe.append(ff)
-                        nc_safe_par.remove(ff)
-
-                if len(nc_protected_par) > 0:
[…]
