### 2023-07-18 pmlmodelling/nctoolkit#81 @robertjwilson
Benchmarking is a tricky one. My experience of this is that it's hard to tell if there is any real difference between nctoolkit and xarray for example.

I've added a notebook I've run on Linux, doing benchmarks of temporal averages and spatial means. Typically, run time is within about 50% of each other. Though xarray tends to be faster for small files, where speed is important.

https://github.com/pmlmodelling/nctoolkit/blob/master/benchmarks/benchmark_nctoolkit_versus_xarray.ipynb

Something like this could be added to the website, but it's hard to do these things without people getting suspicious you are cherry picking and not believing you in the first place.
