### 2023-07-19 pmlmodelling/nctoolkit#84 @robertjwilson
Agreed, this has got bloated lately @malmans2

I've raised an issue to sort this out https://github.com/pmlmodelling/nctoolkit/issues/90.

We run a daily build of nctoolkit on circleci, so I'll need to figure out whether the git-lfs approach or moving to another repo integrates best with our testing set up.
