### 2023-07-19 2192650
Update paper.md with clarifications on plotting functionality
