### 2023-07-19 pmlmodelling/nctoolkit#69 @robertjwilson
The core dependencies are currently listed in the Acknowledgements section of the paper. I have just updated the paper so that it states which key libraries are required for plotting in the Acknowledgements section.
