### 2023-07-24 54ada17
improve colour scale handling
