### 2023-07-24 55060e0
minor speed improvment in para tests
