+++ nctoolkit/cdo_command.py
@@ -2,9 +2,10 @@ import subprocess
 import os
 
 from nctoolkit.runthis import run_this
+from nctoolkit.session import session_info
 
 
-def cdo_command(self, command=None, ensemble=False):
+def cdo_command(self, command=None, ensemble=False, check = False):
     """
     Apply a cdo command
 
@@ -15,6 +16,8 @@ def cdo_command(self, command=None, ensemble=False):
         "cdo {command} infile outfile" will run.
     ensemble : bool
         Is this an ensemble command?
+    check : bool
+        Check whether the command is valid
     """
 
     # First, check that the command is valid
@@ -26,26 +29,45 @@ def cdo_command(self, command=None, ensemble=False):
 
     if command.startswith("cdo "):
         command = command.replace("cdo ", " ").strip()
-    read = subprocess.run(
-        "cdo --operators", shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE
-    ).stdout
 
-    cdo_methods = [x.split(" ")[0].replace("b'", "") for x in str(read).split("\\n")]
+    if check:
+        read = subprocess.run(
+            "cdo --operators", shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+        ).stdout
 
-    cdo_methods = [mm for mm in cdo_methods if len(mm) > 0]
+        cdo_methods = [x.split(" ")[0].replace("b'", "") for x in str(read).split("\\n")]
 
-    cdo_methods.append("L")
-    cdo_methods.append("reduce_dim")
+        cdo_methods = [mm for mm in cdo_methods if len(mm) > 0]
 
-    for x in command.replace("-f ", "").replace("-z ", "").strip().split(" "):
-        y = x.split(",")[0].replace("-", "")
-        if y not in cdo_methods and y != "nc4" and y != "zip_9":
-            if not os.path.exists(y):
-                raise ValueError(f"{y} is not a cdo method!" + y)
+        cdo_methods.append("L")
+        cdo_methods.append("reduce_dim")
+
+        # test whether the command is a valid CDO command
+        test_command = command
+
+        # remove anything within quotes from test_command
+
+        while '"' in test_command:
+            test_command = test_command.replace(
+                test_command[test_command.find('"') : test_command.find('"', 1) + 1], ""
+            )
+
+        while "'" in test_command:
+            test_command = test_command.replace(
+                test_command[test_command.find("'") : test_command.find("'", 1) + 1], ""
+            )
+
+        test_command = test_command.replace("  ", " ")
+
+        for x in test_command.replace("-f ", "").replace("-z ", "").strip().split(" "):
+            y = x.split(",")[0].replace("-", "")
+            if y not in cdo_methods and y != "nc4" and y != "zip_9":
+                if not os.path.exists(y):
+                    raise ValueError(f"{y} is not a cdo method!" + y)
 
     new_command = ""
     for cc in command.split(" "):
-        if cc in cdo_methods:
+        if cc in session_info["cdo_methods"]: 
             new_command += " -" + cc
         else:
             new_command += " " + cc
@@ -62,7 +84,7 @@ def cdo_command(self, command=None, ensemble=False):
         output = "one"
         self._merged = True
 
-    for mm in cdo_methods:
+    for mm in session_info["cdo_methods"]:
         if " " + mm + "," in cdo_command:
             cdo_command = cdo_command.replace(" " + mm + ",", " -" + mm + ",")
     cdo_command = cdo_command.strip()
