### 2023-07-24 935a1d0
switch to cmip6 figure example
