### 2023-07-24 4e3c640
add line on advantages of using cdo
