+++ paper.md
@@ -40,7 +40,7 @@ netCDF is a file format for storing multidimensional data, and it is the fundame
 nctoolkit's core object is the dataset, which is made up of netCDF files stored in a temporary location. Methods use the CDO library to perform operations on the dataset, and methods modify the datasets instead of returning new objects. Evaluation is lazy by default. This means that methods are only evaluated when necessary or when forced, which significantly improves performance. To ensure full functionality of nctoolkit, it is prefable that files analyzed follow the CF conventions [@Hassell2017].
 
 
-The package's core functionality includes the following dataset methods: regridding (`regrid` and `to_latlon`), subsetting (`subset`), temporal statistics (`tmean`, `tmax` etc.), spatial statistics (`spatial_mean`, `spatial_max` etc.), vertical statistics and methods (`vertical_mean`, `vertical_interp`), plotting (`plot`, `pub_plot`), anomaly calculation (`annual_anomaly`), mathematical operations (`assign`) and ensemble statistics (`ensemble_mean` etc.). The package also includes a number of other methods for common tasks, including calculating the difference between two datasets(`ds1-ds2`), extracting the top layer of a dataset (`top`), and calculating the rolling mean (`rolling_mean`). The package also makes it easy to match gridded netCDF data to point observation data using the `match_points` method. 
+The package's core functionality includes the following dataset methods: regridding (`regrid` and `to_latlon`), subsetting (`subset`), temporal statistics (`tmean`, `tmax` etc.), spatial statistics (`spatial_mean`, `spatial_max` etc.), vertical statistics and methods (`vertical_mean`, `vertical_interp`), plotting (`plot`, `pub_plot`), anomaly calculation (`annual_anomaly`), mathematical operations (`assign`) and ensemble statistics (`ensemble_mean` etc.). The package also includes a number of other methods for common tasks, including calculating the difference between two datasets(`ds1-ds2`), extracting the top layer of a dataset (`top`), and calculating the rolling mean (`rolling_mean`). The package also makes it easy to match gridded netCDF data to point observation data using the `match_points` method. Datasets can use multiple files as input, and the `multiprocessing` package is used internally by nctoolkit to enable easy parallelization of operations on multiple files.
 
 
 
