### 2023-07-25 pmlmodelling/nctoolkit#57 @robertjwilson
I've added a statement on parallel processing to the paper. https://github.com/pmlmodelling/nctoolkit/commit/04e09b63d86f6f6f62c8166a1c9825876d5b084a

It's difficult to come up with a defensible performance statement for JOSS without spending days creating rigorous benchmarks, so I'd prefer to avoid it. And it's probably impossible to do given the tight word limit.

As raised here, I've got some benchmarks that suggest similar performance to xarray: https://github.com/pmlmodelling/nctoolkit/issues/81

Though I've yet to test performance on multi-file datasets in xarray versus nctoolkit.
