### 2023-07-25 043cd79
elaborate on netCDF file benefits for non-users
