### 2023-07-25 pmlmodelling/nctoolkit#59 @robertjwilson
I've just added a better statement for non-netCDF users. https://github.com/pmlmodelling/nctoolkit/commit/043cd7976f53fa47019dd6b3369ebe2d45a1eefc

The main issue I'm trying to get across is that netCDF files aren't like csv, and you don't need to reinvent the wheel each time you work with them. This is probably poorly understood outside of groups that work with them all the time.

"Files typically represent spatiotemporal data, such as atmospheric or oceanic temperatures. In contrast to other data formats, such as csv, netCDF files are self-describing and typically follow universally agreed conventions for coordinate names and file structure etc. As a result, it is possible to write software that can work with almost all netCDF files that follow those conventions, and there is no automatic need to burden users with the need to identify the names given to coordinates such as time with the files themselves. A key consequence is that software can carry out operations, such as calculating spatial averages, in one line of code that might otherwise require users to write multiple lines of code, and for these operations to largely work on any netCDF file."

The fact that nctoolkit is more format agnostic than xarray is a plus that could maybe be mentioned explicitly, but I'll maybe avoid that as xarray could easily evolve and make any statement false.
