### 2023-07-25 pmlmodelling/nctoolkit#59 @malmans2
FYI, this xarray accessor might be relevant: https://cf-xarray.readthedocs.io

It decodes CF conventions, so things like `ds.cf.mean(("X", "Y"))` or `ds.cf.mean(("latitude", "longitude"))` work on any dataset, regardless of the arbitrary variable names used.
