### 2023-07-25 pmlmodelling/nctoolkit#68 @robertjwilson
I've created a new example, which shows how to calculate changes in surface temperature from global climate models. This is something that you can just do in nctoolkit, whereas with other tools like xarray or iris, you normally have to google it. The example also uses the `annual_anomaly` method to show that it's easy to calculate changes against baselines.

https://github.com/pmlmodelling/nctoolkit/actions/runs/5654799069

Word limit of 1000 words is tight, so I've added a short list of netCDF packages in the ecosystem.

"The nctoolkit package sits within a Python ecosystem of packages such as xarray and iris, which provide data models and analysis software for netCDF, netCDF4 which provides low level access to netCDF data, and specialist software such as xesmf for processes such as regridding. In contrast to other netCDF libraries, the use of CDO as a back-end allows nctoolkit users to carry out operations, such as spatial averages, without having to specify the specific names of coordinates, such as longitude, latitude and time, which enables code written for one dataset to be easily applied to another."

The key point is that nctoolkit is somewhat more format-agnostic than xarray or iris. Illustrated with the example of `spatial_mean`. In both [xarray](https://docs.xarray.dev/en/stable/examples/area_weighted_temperature.html) and [iris](https://scitools-iris.readthedocs.io/en/latest/generated/gallery/meteorology/plot_COP_1d.html) this requires a few lines of code.

The same is true for temporal averaging in nctoolkit using `tmean` for all types, whereas in xarray it's a slightly confusing combination of `mean`, `groupby` and `resample` to do things. But I'll just stick with spatial average in the paper.
