+++ paper.md
@@ -35,7 +35,7 @@ netCDF is a file format for storing multidimensional data, and it is the fundame
 
 The scale and magnitude of netCDF data in use by scientists continues to grow rapdily. For example, the Coupled Model Intercomparison Project Phase 6 [@ONeill2016], produced approximately 20 PB of publicly available data [@Petrie2021]. This accumulation of data offers great opportunies to environmental scientists, however it also poses challenges because analysis software is often difficult to use by non-specialists [@Bates2018] or is inadequate. nctoolkit is a Python package that aims to fill critical gaps in the current netCDF software ecosystem. It provides a clean interface for working with netCDF files, and it has a particular focus in ensuring the compatibility of methods with oceanic model output, which often have irregular vertical grids. 
 
-The nctoolkit package sits within a Python ecosystem of packages such as `xarray` and `iris`, which provide data models and analysis software for netCDF, `netCDF4` which provides low level access to netCDF data, and specialist software such as `xesmf` for processes such as regridding. In contrast to other netCDF libraries, the use of CDO as a back-end allows nctoolkit users to carry out operations, such as spatial averages, without having to specify the specific names of coordinates, such as longitude, latitude and time, which enables code written for one dataset to be easily applied to another. 
+The nctoolkit package sits within a Python ecosystem of packages such as `xarray` and `iris`, which provide data models and analysis software for netCDF, `netCDF4` which provides low level access to netCDF data. This ecosystem also includes specialist software such as `xesmf` for processes such as regridding and `cf-xarray` which makes `xarray` more format-agnostic. In contrast to other netCDF libraries, the use of CDO as a back-end allows nctoolkit users to carry out operations, such as spatial averages, without having to specify the specific names of coordinates, such as longitude, latitude and time, which enables code written for one dataset to be easily applied to another. 
 
 
 # Overview of Functionality 
