### 2023-07-25 pmlmodelling/nctoolkit#59 @robertjwilson
Hi @malmans2. I had that package in mind when was talking about how xarray could evolve.

I've just added a quick sentence to the paper mentioning cf-xarray.

"This ecosystem also includes specialist software such as xesmf for processes such as regridding and cf-xarray which makes xarray more format-agnostic."
