### 2023-07-25 pmlmodelling/nctoolkit#68 @malmans2
Looks good and I think it makes much more clear the scope of `nctoolkit`.
Could you please add instructions/code to download the example data?
