+++ paper.md
@@ -50,7 +50,9 @@ The package's core functionality includes the following dataset methods: regridd
 
 # Example Use Case
 
-This example shows how to calculate changes in surface temperatures projected by global climate models. The example is from the climate model MPI-ESM-2-LR under the SSP5 8.5 scenario, and we use the r1i1p1f1 variant available from the Earth System Grid Federation. We will show how to map projected changes in temperature between 1850-69 and 2080-99 and also how to calculate the global average temperature change.  The data is stored in multiple netCDF files, which are be opened using the `open_data` function. This returns a `Dataset` object, which contains the data and metadata from the netCDF file. The `Dataset` object has a number of methods for working with the data, which can be used for manipulations and analysis. In this example, we first merge the data along the time dimension. We then use the `annual_anomaly` method to calculate how much temperature will change in each model grid cell. The end of this time series, i.e. the change for 2080-99 is then mapped using `pub_plot`. Finally, we calculate the global average temperature change using the `spatial_mean` method, and plot the time series using `plot`. 
+This example shows how to calculate changes in surface temperatures projected by global climate models. The example is from the climate model MPI-ESM-2-LR under the SSP5 8.5 scenario, and we use the r1i1p1f1 variant. This data is downloadable from the Earth System Grid Federation and is made available on Zenodo: [10.5281/zenodo.8182678](https://zenodo.org/record/8182678).
+ 
+. We will show how to map projected changes in temperature between 1850-69 and 2080-99 and also how to calculate the global average temperature change.  The data is stored in multiple netCDF files, which are be opened using the `open_data` function. This returns a `Dataset` object, which contains the data and metadata from the netCDF file. The `Dataset` object has a number of methods for working with the data, which can be used for manipulations and analysis. In this example, we first merge the data along the time dimension. We then use the `annual_anomaly` method to calculate how much temperature will change in each model grid cell. The end of this time series, i.e. the change for 2080-99 is then mapped using `pub_plot`. Finally, we calculate the global average temperature change using the `spatial_mean` method, and plot the time series using `plot`. 
 
 ```python
 
