### 2023-07-25 c42c8c7
improve land handling in blue-red color scales
