### 2023-07-25 pmlmodelling/nctoolkit#82 @robertjwilson
I've modified the example so it generates directories and the function is now fixed https://nctoolkit.readthedocs.io/en/latest/parallel.html
