### 2023-07-25 e13ec7d
change pub plot test for api change
