### 2023-07-25 298b0f3
add news about version 1.0.0
