### 2023-07-26 #5494 @robertjwilson
I think this all seems to be resolved @martinfleis. I will leave it to @platipodium @malmans2 to close off issues if they are satisfied with everything.

Thank you to both for constructive reviews and the changes made since then have definitely enhanced the package.

Outstanding issues were:

- [ ] Performance. I've added a notebook comparing performance with xarray for key methods. This might go on the website at some point, but my experience is that users are very distrustful of speed claims and just want to see for themselves https://github.com/pmlmodelling/nctoolkit/issues/81
- [ ] Summary. It's now made clearer in the Acknowledgements section that certain dependencies are required for plotting, which is core functionality:  https://github.com/pmlmodelling/nctoolkit/issues/69
- [ ] State of the field. There is now a short section listing key netCDF libraries in Python. I'm happy to add more in here if the reviewers think that's necessary, though the paper is now running against the 1000 word limit. https://github.com/pmlmodelling/nctoolkit/issues/68
- [ ] Community guidelines. These have now been added to the package website: https://nctoolkit.readthedocs.io/en/latest/contributing.html
- [ ] Summary. The paper now has a few sentences outlining more on netCDF files and the advantages of using them.

@malmans2 listed some additional issues that could/should be fixed in future.

We have now modified lazy evaluation, so that it works in a way that won't confuse anyone https://github.com/openjournals/joss-reviews/issues/5494. Version 1.0.0 of nctoolkit has been released this week to contain that change.

The bug in the multiprocessing notebook should now be fixed. https://github.com/pmlmodelling/nctoolkit/issues/82

We are looking into reducing the size of the repo, which we agree is too big. We can't do this immediately as we have daily builds running to test on circleci, so we'll need to figure out how to split out the data and keep out testing workflows active https://github.com/pmlmodelling/nctoolkit/issues/84

We are also looking into how to ensure the documentation is built at runtime.

I believe that is a complete list of the previously outstanding issues. Thanks again to @platipodium and @malmans2 for their good reviews.
