### 2023-07-26 c7b6143
change default diverging scale to RdBu_r
