### 2023-07-26 455d87f
add publication plot to docs
