+++ docs/source/visualization.rst
@@ -68,6 +68,28 @@ In a similar vein, it can automatically handle time series. Below we plot a time
 .. raw:: html
    :file: visualization_plot4.html
 
+Publication quality plots
+---------------------
+
+The ability to produce plots that are at or close to publication quality was introduced in version 0.9.2 in nctoolkit.
+This is carried out using the `pub_plot` method, and is designed to quickly produce a plot that is suitable for a publication or presentation.
+
+.. code:: ipython3
+
+        ff =  "sst.mon.mean.nc"
+        ds = nc.open_data(ff)
+        ds.tmean()
+        ds.pub_plot()
+
+# display a png file called pub_plot1.png
+
+.. image:: ../_static/pub_plot1.png
+   :width: 100%
+   :align: center
+   :alt: pub_plot1
+
+Currently, `pub_plot` is restricted to work with regular lon/lat grids. A limited number of modifications can be made, as changes to the colour scale. Over-time this will become a fully featured method.
+
 
 Plotting internals
 ---------------------
