### 2023-07-26 22bbcb8
fix image path in vis
