### 2023-07-26 f0b3606
drop zero masking
