### 2023-07-27 pmlmodelling/nctoolkit#81 @malmans2
I think some sort of benchmark should be added to the website because both the documentation and GitHub About say: [nctoolkit: Fast and easy analysis of netCDF data in Python](https://nctoolkit.readthedocs.io/en/latest/index.html#nctoolkit-fast-and-easy-analysis-of-netcdf-data-in-python).
Alternatively, you could just use the same title used for the paper:
> `nctoolkit: A Python package for netCDF analysis and post-processing`

About the notebook:

IMO the benchmark notebook is not as clear as the other notebooks.
Here are a few suggestions:
1. Keep it as simple as possible (e.g., can you use [tempfile.NamedTemporaryFile](https://docs.python.org/3/library/tempfile.html#tempfile.NamedTemporaryFile) instead of writing your own code?)
2. Separate xarray and nctoolkit code in dedicated functions
3. Make the results more robust. E.g., if the benchmark is done through jupyter, you could do this: `timing = %timeit func(...)`

Here is an example:
```python
def compare(xr_func, nc_func, check=True):
    results = {}
    for bin in bins:
        # create dataset for benchmark
        with tempfile.NamedTemporaryFile(suffix=".nc") as tmpfile:
            # write dataset for benchmark
            if check:
                xr.testing.assert_allclose(xr_func(tmpfile.name), nc_func(tmpfile.name))
            xr_timing = %timeit xr_func(tmpfile.name)
            nc_timing = %timeit nc_func(tmpfile.name)
            results[bin] = {"xr": xr_timing, "nc": nc_timing}
    return results

def xr_spatial_mean(path):
    ds = xr.open_dataset(path)
    ...
    return ds

def nc_spatial_mean(path):
    ds = nc.open_data(ff_orig)
    ...
    return ds.to_xarray()

spatial_mean_timings = compare(xr_spatial_mean, nc_spatial_mean)
```

One final comment, if dask in not going to be used in the benchmark, I would explicitly mention it as it's an important feature of xarray.
