### 2023-07-27 pmlmodelling/nctoolkit#77 @malmans2
OK, I think it's clear now.

Very minor thing I've noticed. When I run
```python
nc_ds = nc.open_url("https://github.com/pydata/xarray-data/raw/master/air_temperature.nc")
```
there's a weird string printed while downloading. (The string is the second half of the url).
