### 2023-07-27 pmlmodelling/nctoolkit#82 @malmans2
I don't think it's fixed. The new directory has the wrong path.
You are creating the directory `"./new"` and writing to `"../../data/new"`.

As I mentioned in my previous comment, isn't it an issue that the code silently fails?
