### 2023-07-27 pmlmodelling/nctoolkit#80 @malmans2
Did you end up opening an issue? If yes, I would add the reference here.
