### 2023-08-02 2317b23
add invert method
