+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import Command, find_packages, setup
 import sys
 
-DESCRIPTION = "Fast and easy analysis of netCDF data in Python"
+DESCRIPTION = "A Python package for netCDF analysis and post-processing"
 LONG_DESCRIPTION = """
 
 **nctoolkit** is a comprehensive Python (3.6 and above) package for analyzing netCDF data on Linux and macOS.
