### 2023-08-03 pmlmodelling/nctoolkit#77 @robertjwilson
That's strange. This prints OK for me on Linux.

What Python version/OS are you using?

The code is just `print(f"Downloading {x}")`, where x is the string of the url. So It's hard to see what could cause this
