### 2023-08-03 pmlmodelling/nctoolkit#81 @robertjwilson
I've changed the github page title to match the paper title. https://github.com/pmlmodelling/nctoolkit

Originally, the word "fast" was referring to time spent writing code, not really computation. But I'm happy to just replace the title.

In my experience not many people pay attention to or even look at benchmarks. So there is probably little reward to adding it to the website. Certainly xarray and most packages don't have anything.
