### 2023-08-03 pmlmodelling/nctoolkit#82 @robertjwilson
OK. Git screw up. I thought a fix was commited and pushed but it wasn't.

Now definitely fixed https://nctoolkit.readthedocs.io/en/latest/parallel.html

If this is still failing silently can you confirm the OS/Python?
