### 2023-08-03 pmlmodelling/nctoolkit#80 @robertjwilson
I thought I had closed this off. My bad. This should now work fine.

Updates to panel broke ncplot in web browsers in May, and my "fixes" meant files like this weren't plotting. But it should be fine in ncplot v0.3.5.
