### 2023-08-03 pmlmodelling/nctoolkit#81 @malmans2
OK, sorry if I wasn't clear.
> The authors define the software as fast and easy to use. I don't know exactly what fast means in this context (maybe compared to other libraries?), and how to confirm it. The documentation is missing a section that shows that the software is fast.

My point was that it's not clear what fast means in this context, it could mean something like [ruff](https://beta.ruff.rs/docs/) or [this JOSS paper](https://joss.theoj.org/papers/10.21105/joss.03675), which would require a benchmark.

IMO the title you used in the paper is more clear.
If still see the old title both in the About section and in the [latest documentation](https://nctoolkit.readthedocs.io/en/latest/):
![Screenshot 2023-08-03 at 12 36 56](https://github.com/pmlmodelling/nctoolkit/assets/22245117/11aa16fa-44d1-4615-8ac0-a3b5ca7bcc27)
