### 2023-08-03 pmlmodelling/nctoolkit#77 @malmans2
I'm using macOS. Here is the env:
[nctoolkit_env.txt](https://github.com/pmlmodelling/nctoolkit/files/12250099/nctoolkit_env.txt)

I tried both python and ipython, same issue.
