### 2023-08-03 pmlmodelling/nctoolkit#82 @malmans2
The code works now.

Still, if I add a bug on purpose in the function, no Exception is raised (python 3.11, macOS).
It's not relevant for JOSS though, so I'm closing this.
