### 2023-08-03 44b7d06
remove url print in open_url
