+++ nctoolkit/api.py
@@ -606,8 +606,6 @@ def open_data(x=[], checks=True, **kwargs):
             if is_url(x):
                 if thredds is False:
                     new_x = temp_file(".nc")
-                    print(f"Downloading {x}")
-                    print("\033[A                             \033[A")
 
                     if ftp_details is not None and x.startswith("ftp"):
                         user = ftp_details["user"]
