### 2023-08-03 pmlmodelling/nctoolkit#77 @robertjwilson
OK. This seems to be a shell issue. I also have this, which I remember was just to improve the printing.

`print("\033[A                             \033[A")`

This must behave differently on macs.

Printing the url you are downloading is overkill. So I've just removed that from the function in the dev version.
