### 2023-08-03 pmlmodelling/nctoolkit#80 @malmans2
This is now working, so I'm closing the issue.

A couple of things I noticed tho:

- There's a few xarray's deprecation warnings:
```
Warning: xarray could not decode times!
Calling int on a single element Series is deprecated and will raise a TypeError in the future. Use int(ser.iloc[0]) instead
Calling int on a single element Series is deprecated and will raise a TypeError in the future. Use int(ser.iloc[0]) instead
```
- `nc.open_data("~/Downloads/tas_day_GFDL-CM4_historical_r1i1p1f1_gr1_18500101-18691231.nc", checks = False)` did not work, I had to provide the absolute path. I suggest to add support for it.
