### 2023-08-03 #5494 @malmans2
@martinfleis, all my comments have been addressed by the authors.
No further comments from me, I think `nctoolkit` is suitable for publication in JOSS.
