### 2023-08-03 pmlmodelling/nctoolkit#80 @robertjwilson
Thanks. I've just raised an issue. This only impacts macOS, and I've been meaning to fix it for a while. https://github.com/pmlmodelling/nctoolkit/issues/92

I'll double check the deprecation warnings. Last time I looked there was nothing within nctoolkit itself throwing the warnings, it was purely hvplot. So little that could be done in nctoolkit itself.
