### 2023-08-03 45790b8
improve handling of ~ in paths
