### 2023-08-03 pmlmodelling/nctoolkit#57 @platipodium
Fair enough, I think the simple statement you added is sufficient. Thanks!
