### 2023-08-03 pmlmodelling/nctoolkit#58 @platipodium
OK, taking into account the word limit, I think your new paragraph on leveraging CDO as the main feature setting it apart is sufficient.
