### 2023-08-03 #5494 @martinfleis
@platipodium Thanks! Can you tick the remaining boxes in [your checklist](https://github.com/openjournals/joss-reviews/issues/5494#issuecomment-1564851592) above?
