### 2023-08-03 b1128ec
Update paper.bib with matplotlib doi
