+++ paper.bib
@@ -27,6 +27,7 @@ year = {2005}
 @article{hunter2007matplotlib,
   title={Matplotlib: A 2D graphics environment},
   author={Hunter, John D},
+  doi = {10.1109/MCSE.2007.55},
   journal={Computing in science \& engineering},
   volume={9},
   number={03},
