### 2023-08-04 8962fb6
Update affiliations in paper.md
