### 2023-08-04 a76262c
Update paper.bib with mpi-esm2-lr ref
