+++ paper.bib
@@ -13,6 +13,18 @@ volume = {25},
 year = {2005}
 }
 
+@article{mauritsen2019developments,
+  title={Developments in the MPI-M Earth System Model version 1.2 (MPI-ESM1. 2) and its response to increasing CO2},
+  author={Mauritsen, Thorsten and Bader, J{\"u}rgen and Becker, Tobias and Behrens, J{\"o}rg and Bittner, Matthias and Brokopf, Renate and Brovkin, Victor and Claussen, Martin and Crueger, Traute and Esch, Monika and others},
+  journal={Journal of Advances in Modeling Earth Systems},
+  volume={11},
+  doi = {10.1029/2018MS001400},
+  number={4},
+  pages={998--1038},
+  year={2019},
+  publisher={Wiley Online Library}
+}
+
 @article{mckinney2011pandas,
   title={pandas: a foundational Python library for data analysis and statistics},
   author={McKinney, Wes and others},
