### 2023-08-04 f21847f
Update paper.md with typo fixes. Updated refs.
