### 2023-08-04 0d10741
Update paper.md with final language revisions
