+++ paper.md
@@ -44,7 +44,7 @@ The nctoolkit package sits within a Python ecosystem of packages such as `xarray
 nctoolkit's core object is the dataset, which is made up of netCDF files stored in a temporary location. Methods use the CDO library to perform operations on the dataset, and methods modify the datasets instead of returning new objects. Evaluation is lazy by default. This means that methods are only evaluated when necessary or when forced, which significantly improves performance. To ensure full functionality of nctoolkit, it is preferable that files follow the CF conventions [@Hassell2017].
 
 
-The package's core functionality includes the following dataset methods: regridding (`regrid` and `to_latlon`), subsetting (`subset`), temporal statistics (`tmean`, `tmax` etc.), spatial statistics (`spatial_mean`, `spatial_max` etc.), vertical statistics and methods (`vertical_mean`, `vertical_interp`), plotting (`plot`, `pub_plot`), anomaly calculation (`annual_anomaly`), mathematical operations (`assign`) and ensemble statistics (`ensemble_mean` etc.). The package also includes a number of other methods for common tasks, including calculating the difference between two datasets (`ds1-ds2`), extracting the top layer of a dataset (`top`), and calculating the rolling mean (`rolling_mean`). The package also makes it easy to match gridded netCDF data to point observation data using the `match_points` method. Datasets can use multiple files as input, and the `multiprocessing` package is used internally by nctoolkit to enable easy parallelization of operations on multiple files.
+The package's core functionality includes the following dataset methods: regridding (`regrid` and `to_latlon`), subsetting (`subset`), temporal statistics (`tmean`, `tmax` etc.), spatial statistics (`spatial_mean`, `spatial_max` etc.), vertical statistics and methods (`vertical_mean`, `vertical_interp`), plotting (`plot`, `pub_plot`), anomaly calculation (`annual_anomaly`), mathematical operations (`assign`) and ensemble statistics (`ensemble_mean` etc.). The package also includes a number of other methods for common tasks, including calculating the difference between two datasets (`ds1-ds2`), extracting the top and bottom layers of a dataset (`top` and `bottom`), and calculating the rolling mean (`rolling_mean`). The package also makes it easy to match gridded netCDF data to point observation data using the `match_points` method. Datasets can use multiple files as input, and the `multiprocessing` package is used internally by nctoolkit to enable easy parallelization of operations on multiple files.
 
 
 
@@ -52,7 +52,7 @@ The package's core functionality includes the following dataset methods: regridd
 
 This example shows how to calculate changes in surface temperatures projected by global climate models. The example is from the climate model MPI-ESM-2-LR [@mauritsen2019developments] under the SSP5 8.5 scenario, and we use the r1i1p1f1 variant. This data is downloadable from the Earth System Grid Federation and is made available on Zenodo: [10.5281/zenodo.8182678](https://zenodo.org/record/8182678).
  
-We will show how to map projected changes in temperature between 1850-69 and 2080-99 and also how to calculate the global average temperature change.  The data is stored in multiple netCDF files, which are opened using the `open_data` function. This returns a `Dataset` object, which contains the data and metadata from the netCDF file. The `Dataset` object has a number of methods for working with the data, which can be used for manipulations and analysis. In this example, we first merge the data along the time dimension. We then use the `annual_anomaly` method to calculate how much temperature will change in each model grid cell. The end of this time series, i.e. the change for 2080-99 is then mapped using `pub_plot`. Finally, we calculate the global average temperature change using the `spatial_mean` method, and plot the time series using `plot`. 
[…]
