### 2023-08-04 #5494 @robertjwilson
@martinfleis Everything should be done now. I've fixed some typos in the paper and added appropriate refs/acknowledgement for the CMIP6 model used in the example code.

The version number is 1.0.0.

The DOI of the archived version is 10.5281/zenodo.8183367
