### 2023-08-04 2ab1394
correct docstring for set
