+++ nctoolkit/set.py
@@ -1,7 +1,7 @@
 
 def set(self, **kwargs):
     """
-    A method for subsetting datasets to specific variables, years, longitudes etc.
+    A method for setting the units, names, and long names of the dataset.
     Operations are applied in the order supplied.
 
     Parameters
