### 2023-08-04 aaba29e
ensure strip_vars handles tempfiles efficiently
