### 2023-08-04 e4111b0
don't print url when downloading with open_url
