### 2023-08-04 231e148
ensure strip_vars handles tempfiles efficiently
