### 2023-08-04 46e1526
remove some redudant code
