+++ nctoolkit/deprecated.py
@@ -1,8 +1,0 @@
-import pandas as pd
-import subprocess
-import warnings
-
-from datetime import datetime
-
-from nctoolkit.runthis import run_this
-from nctoolkit.session import session_info
