### 2023-08-07 d0498d7
revert to pygments 2.7.4
