+++ docs/requirements.txt
@@ -4,7 +4,7 @@ numpydoc==0.9.2
 #ipykernel==5.1.2
 ipykernel
 nbsphinx>=0.8.6
-Pygments==2.15.0
+Pygments==2.7.4
 sphinx_rtd_theme==0.5.2
 #docutils<0.17
 docutils
