### 2023-08-07 bb396d4
ensure to_nc can handle ~
