### 2023-08-08 b90a56f
fix some typos and spelling etc.
