+++ paper.md
@@ -26,33 +26,33 @@ bibliography: paper.bib
 
 # Summary
 
-nctoolkit is a Python package for the analysis and post-processing of netCDF files. It provides a simple, intuitive interface, and includes a number of functions for common tasks such as subsetting, regridding, statistical analysis and plotting. The package is designed to be easy to use, and to require minimal code to perform common tasks. It is built on top of the Climate Data Operators (CDO)[@schulzweida_uwe_2022_7112925] library, which provides a powerful data model for working with multidimensional data. The underlying goal of the package is to provide comprehensive methods to climate, marine and atmospheric scientists working with netCDF data that will meet 80-100% of their typical processing requirements. 
+nctoolkit is a Python package for the analysis and post-processing of netCDF files. It provides a simple, intuitive interface, and includes a number of functions for common tasks such as subsetting, regridding, statistical analysis and plotting. The package is designed to be easy to use, and to require minimal code for performing common tasks. It is built on top of the Climate Data Operators (CDO) library [@schulzweida_uwe_2022_7112925], which provides a powerful data model for working with multidimensional data. The core aim of the package is to provide over 80% of the typical data processing requirements for climate, marine and atmospheric scientists who work with netCDF data. 
 
 
 # Statement of need
 
-netCDF is a file format for storing multidimensional data, and it is the fundamental storage unit of most modelling and large-scale observational work carried out in climate, marine and atmospheric sciences. Files typically represent spatiotemporal data, such as atmospheric or oceanic temperatures. In contrast to other data formats, such as csv, netCDF files are self-describing and typically follow universally agreed conventions for coordinate names and file structure etc. As a result, it is possible to write software that can work with almost all netCDF files that follow those conventions, and there is no automatic need to burden users with the need to identify the names given to coordinates, such as time, within the files themselves. A key consequence is that software can carry out operations, such as calculating spatial averages, in one line of code that might otherwise require users to write multiple lines of code, and for these operations to largely work on any netCDF file.
+netCDF is a file format for storing multidimensional data, and it is the fundamental storage unit of most modelling and large-scale observational work carried out in climate, marine and atmospheric science. Files typically represent spatiotemporal data, such as atmospheric or oceanic temperatures. In contrast to other data formats, such as csv, netCDF files are self-describing and typically follow universally agreed conventions for coordinate names and file structure etc. As a result, it is possible to write software that can work with almost all netCDF files that follow these conventions, and there is no automatic neeed to burden users with the task of identifying the names given to coordinates, such as time, within the files themselves. Software can therefore be written that will carry out operations, such as calculating spatial averages, in one line of code that might otherwise require users to write multiple lines of code, and for these operations to largely work on any netCDF file.
 
-The scale and magnitude of netCDF data in use by scientists continues to grow rapdily. For example, the Coupled Model Intercomparison Project Phase 6 [@ONeill2016], produced approximately 20 PB of publicly available data [@Petrie2021]. This accumulation of data offers great opportunies to environmental scientists. However it also poses challenges because analysis software is often difficult to use by non-specialists [@Bates2018] or is inadequate. nctoolkit is a Python package that aims to fill critical gaps in the current netCDF software ecosystem. It provides a clean interface for working with netCDF files, and it has a particular focus in ensuring the compatibility of methods with oceanic model output, which often have irregular vertical grids. 
[…]
