+++ paper.bib
@@ -13,7 +13,7 @@ year = {2005}
 }
 
 @article{mauritsen2019developments,
-  title={Developments in the MPI-M Earth System Model version 1.2 (MPI-ESM1. 2) and its response to increasing CO2},
+  title={{Developments in the MPI-M Earth System Model version 1.2 (MPI-ESM1. 2) and its response to increasing CO2}},
   author={Mauritsen, Thorsten and Bader, J{\"u}rgen and Becker, Tobias and Behrens, J{\"o}rg and Bittner, Matthias and Brokopf, Renate and Brovkin, Victor and Claussen, Martin and Crueger, Traute and Esch, Monika et al.},
   journal={Journal of Advances in Modeling Earth Systems},
   volume={11},
@@ -25,7 +25,7 @@ year = {2005}
 }
 
 @article{mckinney2011pandas,
-  title={pandas: a foundational Python library for data analysis and statistics},
+  title={{pandas: a foundational Python library for data analysis and statistics}},
   author={McKinney, Wes},
   journal={Python for high performance and scientific computing},
   volume={14},
@@ -36,7 +36,7 @@ year = {2005}
 }
 
 @article{hunter2007matplotlib,
-  title={Matplotlib: A 2D graphics environment},
+  title={{Matplotlib: A 2D graphics environment}},
   author={Hunter, John D},
   doi = {10.1109/MCSE.2007.55},
   journal={Computing in science \& engineering},
@@ -48,7 +48,7 @@ year = {2005}
 }
 
 @misc{eaton2003netcdf,
-  title={NetCDF Climate and Forecast (CF) metadata conventions},
+  title={{NetCDF Climate and Forecast (CF) metadata conventions}},
   author={Eaton, Brian and Gregory, Jonathan and Drach, Bob and Taylor, Karl and Hankin, Steve and Caron, John and Signell, Rich and Bentley, Phil and Rappa, Greg and H{\"o}ck, Heinke and others},
   year={2003},
   publisher={Version}
@@ -56,7 +56,7 @@ year = {2005}
 
 @misc{schulzweida_uwe_2022_7112925,
   author       = {Schulzweida, Uwe},
-  title        = {CDO User Guide},
+  title        = {{CDO User Guide}},
   month        = oct,
   year         = 2022,
   publisher    = {Zenodo},
