### 2023-08-08 44c4b35
minor tidying of paper/refs
