+++ paper.bib
@@ -1,11 +1,7 @@
 @article{Ishii2005,
-abstract = {Data for the 20th century from the International Comprehensive Ocean and Atmosphere Data Set and the Kobe Collection have been used as input data for global objective analyses of sea-surface temperatures (SSTs) and other marine meteorological variables. This study seeks a better understanding of the historical marine meteorological data and an evaluation of the quality of the data in the Kobe Collection. Objective analyses yield gridded data that are less noisy than observed data, which facilitates handling of historical data. The observed data determine the quality of the objective analyses, and quality control specified for historical data is incorporated into the objective analysis to reduce artificial errors. The objective analyses are based on optimum interpolation and reconstruction with empirical orthogonal functions. The final database produced in this study not only contains analysed values, but also analysis errors and data distributions at each time step of the objective analyses. The analysis database contains amp le information on historical observations, as well as signals of marine climate variations during the century. Time series of global mean marine temperatures and cloud cover include trends linked to global warming, and local peaks appear commonly in all the time series around the 1940s. Sea-level pressure and sea-surface wind fields show significant linear trends at high latitudes and over the North Pacific Ocean respectively. These trends seem to be artificial. An SST analysis used widely in climatological studies was verified against HadISST from the Hadley Centre and an SST analysis derived from satellite and in situ observations. El Ni{\~{n}}o and southern oscillation indices for the century are successfully reproduced, even though observations in the tropics are much rarer before 1950 than after 1950. Copyright {\textcopyright} 2005 Royal Meteorological Society.},
 author = {Ishii, Masayoshi and Shouji, Akiko and Sugimoto, Satoshi and Matsumoto, Takanori},
 doi = {10.1002/joc.1169},
-issn = {08998418},
 journal = {International Journal of Climatology},
-keywords = {ICOADS,Kobe Collection,Marine meteorology,Objective analysis,Optimum interpolation,Reconstruction,SST},
-number = {7},
 pages = {865--879},
 title = {{Objective analyses of sea-surface temperature and marine meteorological variables for the 20th century using ICOADS and the Kobe Collection}},
 volume = {25},
@@ -18,7 +14,6 @@ year = {2005}
   journal={Journal of Advances in Modeling Earth Systems},
   volume={11},
   doi = {10.1029/2018MS001400},
-  number={4},
   pages={998--1038},
   year={2019},
   publisher={Wiley Online Library}
@@ -29,7 +24,6 @@ year = {2005}
   author={McKinney, Wes},
   journal={Python for high performance and scientific computing},
   volume={14},
-  number={9},
   pages={1--9},
   year={2011},
   publisher={Seattle}
@@ -41,7 +35,6 @@ year = {2005}
   doi = {10.1109/MCSE.2007.55},
   journal={Computing in science \& engineering},
   volume={9},
-  number={03},
   pages={90--95},
   year={2007},
   publisher={IEEE Computer Society}
@@ -67,12 +60,9 @@ year = {2005}
 
 
 @article{Hassell2017,
[…]
