+++ paper.bib
@@ -9,7 +9,7 @@ year = {2005}
 }
 
 @article{mauritsen2019developments,
-  title={{Developments in the MPI-M Earth System Model version 1.2 (MPI-ESM1.2) and Its response to increasing CO2}},
+  title={{Developments in the MPI-M Earth System Model version 1.2 (MPI-ESM1.2) and Its Response to Increasing CO\_2}},
   author={Mauritsen, Thorsten and Bader, J{\"u}rgen and Becker, Tobias and Behrens, J{\"o}rg and Bittner, Matthias and Brokopf, Renate and Brovkin, Victor and Claussen, Martin and Crueger, Traute and Esch, Monika et al.},
   journal={Journal of Advances in Modeling Earth Systems},
   volume={11},
