### 2023-08-08 b00508e
fix nco_command ensemble bug
