### 2023-08-08 6aea55b
Update paper.md with a couple of typo fixes
