### 2023-08-09 #5494 @kthyng
Ok looks good. I'm interested in this package because I find nco tools difficult to use in terms of what command to use — maybe there are too many options!
